% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/borders.R
\name{borders}
\alias{borders}
\title{polygons to borders}
\usage{
borders(spdf, tol = 1, id = NULL)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame}

\item{tol}{tolerance to detect contiguity (in map units)}

\item{id}{unique id on the spdf}
}
\value{
SpatialLinesDataFrame
}
\description{
between polygons
}
\details{
Extract borders from a SpatialPolygonsDataFrame
}
\examples{
data(nuts2006)
plot(nuts0.spdf ,col="#CCCCCC",border="white",lwd=1)
contig <-borders(spdf=nuts0.spdf)
plot(contig,col="red",lwd=2,add=TRUE)
}

