% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carroyage.R
\name{carroyage}
\alias{carroyage}
\title{carroyage}
\usage{
carroyage(spdf, celsize, spdfid = NULL)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame}

\item{celsize}{size of the cell}

\item{spdfid}{unique id on the spdf}
}
\value{
SpatialLinesDataFrame
}
\description{
Build a regular grid
}
\details{
Build a regular grid from a SpatialPolygonsDataFrame and a table of surface intersection
}
\examples{
data(nuts2006)
mygrid<-carroyage(spdf=nuts2.spdf,celsize=200000)
plot(mygrid$spdf)
head(mygrid$df)
}

