% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{carto.pal.info}
\alias{carto.pal.info}
\title{Display the Names of all Cartographic Palettes}
\usage{
carto.pal.info()
}
\value{
A vector of palettes names is returned.
}
\description{
Display the names of all the available color palettes.
}
\examples{
carto.pal.info
}
\seealso{
\link{carto.pal}, \link{display.carto.pal}, \link{display.carto.all}
}

