% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{cartography.colors}
\alias{cartography.colors}
\title{Color Palettes}
\format{list}
\source{
UMS RIATE
}
\description{
Color gradients adapted to thematic cartography
}
\details{
xxx
}

