% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{cartography}
\alias{cartography}
\alias{cartography-package}
\title{Cartography Package}
\description{
This package proposes various usefull functions in the map creation process.
It allows various cartographic representation :
\itemize{
\item{Proportionnal symbols maps (circles, squares, bars)}
\item{Chroropleth maps (main discretization methods are availables)}
\item{Flow maps (proportionnal and classified links)}
\item{Discontinuities maps (variable size and color of borders)}
}

It also proposes some additional usefull features like:
\itemize{
\item{Cartographic palettes (palettes adapted to cartographic representation)}
\item{Layout (scale, nort arrow, title...)}
\item{Nice legends}
\item{Access to cartographic API (via OpenStreetMap package)}
\item{Irregular polygons to regular grid transformation with data handling}
}
}

