% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internL_choro.R
\name{choro}
\alias{choro}
\title{choro}
\usage{
choro(var, distr = NULL, col = NULL, nbclass = NULL,
  method = "quantile")
}
\arguments{
\item{var}{vector of valuesused}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null, the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal", "quantile", "jenks","q6","geom")}
}
\value{
List: a vector of colors, colors and distr
}
\description{
add color gradients to spdf according to data classes
}

