% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{coasts.spdf}
\alias{coasts.spdf}
\title{Coastline}
\format{This basemap is a SpatialLinesDataFrame.}
\source{
UMS RIATE
}
\description{
Coastline
}
\details{
The projection used is Lambert azimuthal equal-area projection
(EPSG:3035)
}

