% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.carto.all}
\alias{display.carto.all}
\title{Display All Cartographic Palettes}
\usage{
display.carto.all(n = 10)
}
\arguments{
\item{n}{numeric; number of colors by palettes (from 1 to 20).}
}
\value{
A plot is returned.
}
\description{
Display all the available color palettes.
}
\examples{
display.carto.all(1)
display.carto.all(5)
display.carto.all(8)
display.carto.all(12)
display.carto.all(20)
}

