% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.carto.pal}
\alias{display.carto.pal}
\title{Display one Cartographic Palette}
\usage{
display.carto.pal(name)
}
\arguments{
\item{name}{name of the palette available in the package (see Details).}
}
\description{
Display one color palette.
}
\details{
Available palettes are: "blue.pal", "orange.pal", "red.pal", "brown.pal", "green.pal", "purple.pal"
"pink.pal", "wine.pal", "grey.pal", "turquoise.pal", "sand.pal", "taupe.pal", "kaki.pal" , "harmo.pal".
}
\examples{
display.carto.pal("orange.pal")
display.carto.pal("sand.pal")
}
\seealso{
\link{carto.pal}, \link{display.carto.all}
}

