% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{frame.spdf}
\alias{frame.spdf}
\title{Frame}
\format{This basemap is a SpatialPolygonsDataFrame.}
\source{
UMS RIATE
}
\description{
Frame
}
\details{
The projection used is Lambert azimuthal equal-area projection
(EPSG:3035)
}

