% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{FRdep}
\alias{FRdep}
\title{French Departments Dataset}
\format{This data.frame has 96 rows and 80 columns.}
\source{
Institut national de la statistique et des études économiques
(INSEE). General census of population 1975, 1982, 1990, 1999, 2006, 2011.
Data gathered on November 2014.
}
\usage{
data(FRdep)
}
\description{
This dataset contains some socio-economic data on french
metropolitan departments from 1975 to 2011 (population, age structure,
employement, sector of employement).
}
\details{
This data.frame can be joined to the \code{\link{FRdep.spdf}}
SpatialPolygonsDataframe. The \code{dep_code} field from FRdep matches
the \code{dep_code} field from FRdep.spdf.
"\code{****}" in field names refers to the census years (1975, 1982, 1990,
1999, 2006, 2011).
}
\section{Fields}{

\describe{
\item{\code{dep_code}}{department code (character)}

\item{\code{dep_name}}{department name (character)}

\item{\code{pXXXX_****}}{Population (numeric). "\code{XXXX}" refers to the
different age groups.
\itemize{
 \item{p0014: 0-14 years}
 \item{p1524: 15-24 years}
 \item{p2549: 25-49 years}
 \item{p5064: 50-64 years}
 \item{p6579: 65-79 years}
 \item{p80ol: 80 years and older}
 \item{ptota: total population}
 }}

\item{\code{empl_****}}{25-54 years old active population in employment (numeric).}

\item{\code{unem_****}}{25-54 years old unemployed population(numeric).}

\item{\code{farm_****}}{25-54 years old population employed in the farming sector
(numeric).}

\item{\code{indu_****}}{25-54 years old population employed in the industrial sector
(numeric).}

\item{\code{cons_****}}{25-54 years old population employed in the construction
sector
(numeric).}

\item{\code{serv_****}}{25-54 years old population employed in the service sector
(numeric).}
}}
\seealso{
\code{\link{FRdep.spdf}}
}

