% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{FRdep.spdf}
\alias{FRdep.spdf}
\title{French Departments Dataset}
\format{This basemap is a SpatialPolygonsDataFrame.}
\source{
Geofla départements 2013, IGN.
}
\usage{
data(FRdep)
}
\description{
Basemap of the french departments
}
\details{
This data.frame can be joined to the \code{\link{FRdep.spdf}}
SpatialPolygonsDataframe. The \code{dep_code} field from FRdep matches
the \code{dep_code} field from FRdep.spdf.
The projection used is RGF93 / Lambert-93 (EPSG:2154).
}
\section{Fields}{

\describe{
\item{\code{dep_code}}{department code (character)}
}}
\seealso{
\code{\link{FRdep}}
}

