% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telsncf.R
\name{get_sncf}
\alias{get_sncf}
\title{fonction pour telecharger un ficher de api sncf}
\usage{
get_sncf(dataset = "ponctualite-mensuelle-transilien")
}
\arguments{
\item{dataset}{nom de la dataset}
}
\value{
a \code{data.frame}
}
\description{
fonction pour telecharger un ficher de api sncf
}
\examples{
\dontrun{
get_sncf("ponctualite-mensuelle-transilien")
get_sncf ("tarif-abonnement-ile-de-france")
}
}
