% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridLayer.R
\name{getGridData}
\alias{getGridData}
\title{Compute Data For A Grid Layer}
\usage{
getGridData(x, df, dfid = NULL, var)
}
\arguments{
\item{x}{list; list generated by the \link{gridLayer} function.}

\item{df}{data.frame; \code{df} contains the values to adapt to the grid.}

\item{dfid}{character; id field in \code{df}, default to the first column
of \code{df}. (optional)}

\item{var}{character; name of the numeric field in \code{df} to adapt to the grid.}
}
\value{
A data.frame is returned.
}
\description{
This function compute the data within for grid layer according to surface intersection.
}
\details{
We have to describe precisely the function here
}

