% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridLayer.R
\name{getGridLayer}
\alias{getGridLayer}
\title{Build a Regular Grid Layer}
\usage{
getGridLayer(x, id = NULL, cellsize, type = "regular")
}
\arguments{
\item{x}{an sf object, a simple feature collection (or a SpatialPolygonsDataFrame).}

\item{id}{identifier field in x, default to the first column. (optional)}

\item{cellsize}{targeted area of the cell, in map units.}

\item{type}{shape of the cell, "regular" for squares, "hexagonal" for hexagons.}
}
\value{
A list is returned. The list contains "grid": a polygon sf object of 
a regular grid and "df":  a data frame of surface intersection. df fields are 
id_cell: ids of the grid; id_geo: ids of x objects and area_pct: share of the 
area of the polygon in the cell (a value of 55 means that 55\% of the sf unit area 
is within the cell).
}
\description{
Build a regular grid based on an sf object or a SpatialPolygonsDataFrame. 
Provide also a table of surface intersections.
}
\examples{
\dontrun{
data(nuts2006)
# Get a grid layer
mygrid <- getGridLayer(spdf = nuts2.spdf, cellsize = 200000)
# Plot the grid
plot(mygrid$spdf)
head(mygrid$df)
       }
}
\seealso{
\link{getGridData}
}
