% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLinkLayer.R
\name{getLinkLayer}
\alias{getLinkLayer}
\title{Create a SpatialLinesDataFrame from a data.frame of Links.}
\usage{
getLinkLayer(spdf, spdf2 = NULL, df, spdfid = NULL, spdf2id = NULL,
  dfids = NULL, dfide = NULL)
}
\arguments{
\item{spdf}{SpatialPointsDataFrame or SpatialPolygonsDataFrame; layer used to
get starting points of links. If spdf2 is NULL, spdf is also used to get ending points.
If spdf is a SpatialPolygonsDataFrame points start (or end) at centroids.}

\item{spdf2}{SpatialPointsDataFrame or SpatialPolygonsDataFrame; layer used to
get ending points of links.
If spdf is a SpatialPolygonsDataFrame points start (or end) at centroids.}

\item{df}{data.frame; df contains identifiers of starting and ending points.}

\item{spdfid}{character; id field in spdf, default to the first column
of the spdf data.frame. (optional)}

\item{spdf2id}{character; id field in spdf2, default to the first column
of the spdf2 data.frame. (optional)}

\item{dfids}{character; id field of starting points of links in df, default to the first column
of df. (optional)}

\item{dfide}{character; id field of ending points of links in df, default to the second column
of df. (optional)}
}
\value{
A SpatialLinesDataFrame is returned, its data.frame contains two fields (dfid1 and dfid2).
}
\description{
Create a SpatialLinesDataFrame from a data.frame of links.
}
\examples{
data("nuts2006")
# Create a link layer
twincitiesHU <- twincities[substr(twincities$i,1,2)=="IE", ]
twincities.spdf <- getLinkLayer(spdf = nuts2.spdf, df = twincitiesHU[,1:2])
# plot the links
plot(nuts2.spdf, col = "#6C6870")
plot(twincities.spdf, col = "#F78194", add = TRUE)
}

