% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmLayer.R
\name{getOSMLayer}
\alias{getOSMLayer}
\title{Get an OpenStreetMap Object}
\usage{
getOSMLayer(spdf, type = "osm", zoom = NULL)
}
\arguments{
\item{spdf}{Spatial*DataFrame with projection information}

\item{type}{Type of map used in the \code{OpenStreetMap} package}

\item{zoom}{Zoom level of the map}
}
\description{
Get an OpenStreetMap Object
}
\details{
This function uses rgdal and OpenStreetMap packages.
}
\examples{
data("TNdeleg")
}

