% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{graticule.spdf}
\alias{graticule.spdf}
\title{Graticule}
\format{This basemap is a SpatialLinesDataFrame.}
\source{
UMS RIATE
}
\description{
Graticule
}
\details{
The projection used is Lambert azimuthal equal-area projection
(EPSG:3035)
}

