% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridLayer.R
\name{gridLayer}
\alias{getGridLayer}
\alias{gridLayer}
\title{Build A Regular Grid Layer}
\usage{
getGridLayer(spdf, cellsize, spdfid = NULL)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame.}

\item{cellsize}{numeric; cell size in map units.}

\item{spdfid}{character; id field in \code{spdf}, default to the first column
of the \code{spdf} data.frame. (optional)}
}
\value{
A regular grid and a table of surface intersection are returned.
}
\description{
Build a regular grid based on a SpatialPolygonsDataFrame.
Provide also a table of surface intersections.
}
\details{
We have to describe precisely the function here
}

