% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rempl.R
\name{imput_na}
\alias{imput_na}
\title{remplacement de valeurs manquantes}
\usage{
imput_na(x, method = c("moyenne", "mediane"))
}
\arguments{
\item{x}{numeric vector}

\item{method}{choix entre moyenne et mediane}
}
\value{
a numeric
}
\description{
remplacement de valeurs manquantes
}
\examples{
imput_na(c(1,2,NA,6,7)) #
}
