% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelLayer.R
\name{labelLayer}
\alias{labelLayer}
\title{labelLayer function.}
\usage{
labelLayer(spdf, df, spdfid = NA, dfid = NA, txt, col = "black",
  cex = 0.7, ...)
}
\arguments{
\item{spdf}{Spatial*DataFrame}

\item{df}{DataFrame with Ids and Labels}

\item{spdfid}{Ids of the obj Spatial*DataFrame}

\item{dfid}{Ids of the DataFrame}

\item{txt}{Labels field in data}

\item{col}{Labels color}

\item{cex}{Labels size}

\item{...}{further \link{text} arguments, such as \code{pos} or \code{adj}.}
}
\description{
Put labels
}
\examples{
data("TNdeleg")
}

