% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelLayer.R
\name{labelLayer2}
\alias{labelLayer2}
\title{Title}
\usage{
labelLayer2(x, spdf, df, spdfid = NULL, dfid = NULL, txt, col = "black",
  cex = 0.7, overlap = TRUE, show.lines = TRUE, halo = FALSE,
  bg = "white", r = 0.1, ...)
}
\arguments{
\item{x}{t}

\item{spdf}{c}

\item{df}{c}

\item{spdfid}{c}

\item{dfid}{c}

\item{txt}{c}

\item{col}{c}

\item{cex}{t}

\item{overlap}{df}

\item{show.lines}{t}

\item{halo}{fd}

\item{bg}{d}

\item{r}{d}

\item{...}{t}
}
\value{
dgjsdfj
}
\description{
Title
}
\examples{
mtq <- st_read(system.file("shape/martinique.shp", package="cartography"))
plot(st_geometry(mtq), col = "olivedrab1", border = "olivedrab4", 
     bg = "lightsteelblue3")
labelLayer2(x = mtq, txt = "LIBGEO")
}
