% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legends.R
\name{legendBarsSymbols}
\alias{legendBarsSymbols}
\title{Legend For Proportionnal Bars Maps}
\usage{
legendBarsSymbols(pos = "topleft", title.txt = "Title of the legend",
  title.cex = 0.8, cex = 1, values.cex = 0.6, var, r, breakval = NULL,
  col = "red", col2 = "blue", frame = FALSE, values.rnd = 0,
  style = "c")
}
\arguments{
\item{pos}{character; position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{title.txt}{character; title of the legend.}

\item{title.cex}{numeric; size of the legend title.}

\item{cex}{character; size of the legend. 2 means two times bigger.}

\item{values.cex}{numeric; size of the values in the legend.}

\item{var}{numeric; vector of values.}

\item{r}{numeric; a vector giving the heights of the bars.}

\item{breakval}{numeric; breaking value (see Details).}

\item{col}{character; color of symbols.}

\item{col2}{character; second color of symbols (see Details).}

\item{frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{values.rnd}{numeric; number of decimal places of the values in
the legend.}

\item{style}{character; either "c" or "e". The legend has two display
styles, "c" stands for compact and "e" for extended.}
}
\description{
Plot legend for proportional bars maps
}
\details{
The breakval parameter allows to plot symbols of two
colors: the first color (col) for values superior or equal to breakval,
second color (col2) for values inferior to breakval.
}
\examples{
data("nuts2006")
plot(nuts0.spdf)

legendBarsSymbols(pos = "topleft", title.txt = "Title of\\nthe legend",
                     title.cex = 0.8, values.cex = 0.6,cex = 3,
                     var = nuts1.df$pop2008,
                     r = sqrt((abs(nuts1.df$pop2008) * 100000) / pi),
                     col = "purple",
                     values.rnd=0, style ="e")
}

