% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legends.R
\name{LegendChineseSymbols}
\alias{LegendChineseSymbols}
\title{LegendChineseSymbols}
\usage{
LegendChineseSymbols(pos = "topleft", legTitle, var1label, var2label,
  legTitleCex = 0.8, legValuesCex = 0.6, vect1, vect2, sizevect1, sizevect2,
  col1 = "red", col2 = "blue", frame = FALSE, round = 0, type = "a")
}
\arguments{
\item{pos}{bottomleft, topleft, topright, bottomright, left, right, top, bottom, middle}

\item{legTitle}{Title of the legend}

\item{var1label}{label 1}

\item{var2label}{label 1}

\item{legTitleCex}{legTitleCex}

\item{legValuesCex}{legValuesCex}

\item{vect1}{vect1}

\item{vect2}{vect2}

\item{sizevect1}{size 1}

\item{sizevect2}{size2}

\item{col1}{color defalul = "red"}

\item{col2}{color default = "blue"}

\item{frame}{if TRUE, a frame is drawn}

\item{round}{round}

\item{type}{Type of display of the legend. "a" or "b"}
}
\value{
plot
}
\description{
legend for Chinese Hats layers
}
\details{
plot a legend
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
rect(par()$usr[1], par()$usr[3], par()$usr[2], par()$usr[4], border = "black")
vect1 <- round((nuts0.df$pop2008 / sum(nuts0.df$pop2008))*100,2)
vect2 <- round((nuts0.df$gdppps2008 / sum(nuts0.df$gdppps2008))*100,2)
sizevect1 <- sqrt(vect1)/2*100000
sizevect2 <- sqrt(vect2)/2*100000
LegendChineseSymbols(pos = "topright", var1label = "population totale (habs)",
                    var2label = "pib (euros)", legTitle="PIB par habitant",
                      legTitleCex = 0.8, legValuesCex = 0.6,
                      vect1 = vect1, vect2 = vect2, sizevect1 = sizevect1, sizevect2 = sizevect2,
                      col1="green", col2="yellow", frame=TRUE, round=2, type="b")

LegendChineseSymbols(pos = "bottomleft", var1label = "population totale (habs)",
                      var2label = "pib (euros)", legTitle="PIB par habitant",
                      legTitleCex = 0.8, legValuesCex = 0.6,
                      vect1 = vect1, vect2 = vect2, sizevect1 = sizevect1, sizevect2 = sizevect2,
                      col1="green", col2="yellow", frame=TRUE, round=2, type="a")
}

