% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legends.R
\name{LegendHeightSymbols}
\alias{LegendHeightSymbols}
\title{LegendHeightSymbols}
\usage{
LegendHeightSymbols(pos = "topleft", legTitle = "Title of the legend",
  legTitleCex = 0.8, legValuesCex = 0.6, varvect, sizevect, breakval,
  col1 = "red", col2 = "blue", frame = FALSE, round = 0, type = "a")
}
\arguments{
\item{pos}{bottomleft, topleft, topright, bottomright, left, right, top, bottom, middle}

\item{legTitle}{Title of the legend}

\item{legTitleCex}{legTitleCex}

\item{legValuesCex}{legValuesCex}

\item{varvect}{Values}

\item{sizevect}{Height of the symbols}

\item{breakval}{breakval}

\item{col1}{color defalul = "red"}

\item{col2}{color default = "blue"}

\item{frame}{if TRUE, a frame is drawn}

\item{round}{round}

\item{type}{Type of display of the legend. "a" or "b"}
}
\value{
plot
}
\description{
legend for Prop PropSymbols layers (height)
}
\details{
plot a legend
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
rect(par()$usr[1], par()$usr[3], par()$usr[2], par()$usr[4], border = "black")
LegendHeightSymbols(pos = "topright", legTitle = "Title of the legend",
                    legTitleCex = 0.8, legValuesCex = 0.6,
                    varvect = nuts1.df$pop2008,
                    sizevect = sqrt((abs(nuts1.df$pop2008) * 1000000) / pi),
                    breakval=1000, col1="red", col2="blue", frame=TRUE, round=0, type ="a")

LegendHeightSymbols(pos = "left", legTitle = "Title of the legend",
                    legTitleCex = 0.8, legValuesCex = 0.6,
                    varvect = nuts1.df$pop2008,
                    sizevect = sqrt((abs(nuts1.df$pop2008) * 600) / pi),
                    breakval=NULL, col1="yellow", col2="blue", frame=FALSE, round=0, type ="b")
}

