% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legends.R
\name{legendPropLines}
\alias{legendPropLines}
\title{Legend For Proportionnal Lines Maps}
\usage{
legendPropLines(pos = "topleft", title.txt = "Title of the legend",
  title.cex = 0.8, cex = 1, values.cex = 0.6, var, lwd, col = "red",
  frame = FALSE, values.rnd = 0)
}
\arguments{
\item{pos}{character; position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{title.txt}{character; title of the legend.}

\item{title.cex}{numeric; size of the legend title.}

\item{cex}{character; size of the legend. 2 means two times bigger.}

\item{values.cex}{numeric; size of the values in the legend.}

\item{var}{numeric; vector of values.}

\item{lwd}{numeric; a vector giving the width of the lines.}

\item{col}{character; color of symbols.}

\item{frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{values.rnd}{numeric; number of decimal places of the values in
the legend.}
}
\description{
Plot legend for proportionnal lines maps
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
box()
legendPropLines(pos = "topleft", title.txt = "Title",
                title.cex = 0.8, values.cex = 0.6, cex = 2,
                var = nuts1.df$pop2008,
                lwd = nuts1.df$pop2008/1000000,
                col="red", frame=TRUE, values.rnd=2)
}

