% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legends.R
\name{LegendSizeLines}
\alias{LegendSizeLines}
\title{LegendSizeLines}
\usage{
LegendSizeLines(pos = "topleft", legTitle = "Title of the legend",
  legTitleCex = 0.8, legValuesCex = 0.6, distr, thickness, col, round = 2,
  nodata = TRUE, nodatalabel = "No data", frame = FALSE)
}
\arguments{
\item{pos}{bottomleft, topleft, topright, bottomright, left, right, top, bottom, middle}

\item{legTitle}{Title of the legend}

\item{legTitleCex}{legTitleCex}

\item{legValuesCex}{legTitleCex}

\item{distr}{vector of classes}

\item{thickness}{vector of thickness}

\item{col}{color defalul = "red"}

\item{round}{round}

\item{nodata}{if TRUE, a box "no data" is drawn}

\item{nodatalabel}{label for no data.}

\item{frame}{if TRUE, a frame is drawn}
}
\value{
plot
}
\description{
legend for classes of  Lines sizes layers
}
\details{
plot a legend
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
LegendSizeLines(pos = "bottomleft", legTitle = "Title of the legend", legTitleCex = 0.8,
                legValuesCex = 0.6, distr = c(1,2,3,4,10.2,15.2),
                thickness = c(0.2,2,5,10,27),
                col ="blue", round =2,
                nodata = TRUE,
                nodatalabel = "No data available", frame = TRUE)
}

