% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legends.R
\name{legendText}
\alias{legendText}
\title{Legend To display only a text}
\usage{
legendText(pos = "topleft", txt = "Title of the legend", cex = 0.6,
  col = "black", frame = FALSE)
}
\arguments{
\item{pos}{character; position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{txt}{character; text to be displayed.}

\item{cex}{numeric; size of the legend text}

\item{col}{character; color of the text.}

\item{frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}
}
\description{
Plot legend with a simple textbox
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
box()
legendText(pos = "topleft",txt = "Title of the legend", cex = 0.6,col="black",frame=FALSE)
}

