% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphist.R
\name{maphist}
\alias{maphist}
\title{Histogram on maps}
\usage{
maphist(
  v,
  breaks = NULL,
  method = "quantile",
  nclass = NULL,
  col = NULL,
  pos = "topleft",
  hist.width = 0.25,
  hist.height = 0.35,
  title.txt = NULL,
  title.cex = 0.8,
  axes = TRUE,
  axes.cex = 0.6,
  frame = FALSE
)
}
\arguments{
\item{v}{a vector of numeric values.}

\item{breaks, method, nclass}{classification options, see \link{getBreaks}.}

\item{col}{a vector of colors.}

\item{pos}{position of the legend, one of "topleft", "top", 
"topright", "right", "bottomright", "bottom", "bottomleft", 
"bottomleftextra", "left" or a vector of two coordinates in map units 
(c(x, y)).}

\item{hist.width, hist.height}{width and height of the histogram, in normalized device coordinates (NDC). See Details.}

\item{title.txt}{title of the legend.}

\item{title.cex}{size of the legend title.}

\item{axes}{whether to add a horizontal (\code{"h"}), vertical (\code{"v"}) or both (\code{TRUE}) axes.}

\item{axes.cex}{size of the values in the axes.}

\item{frame}{whether to add a frame to the legend (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Plot an histogram over a map.
}
\details{
\code{hist.width} and \code{hist.height} are defined in NDC, meaning that
the height and width are percentages of the overall dimensions of the device plot. Default parameters produce
a histogram that covers 25\% (on width) and 35\% (on height) of the overall device plot.
}
\examples{
library(sf)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package="cartography"))
# Population density
mtq$POPDENS <- 1e6 * mtq$POP / st_area(x = mtq)

choroLayer(x = mtq, var = "POPDENS",
          border = "grey40",
          legend.pos = "topright", legend.values.rnd = 0,
          legend.title.txt = "Population Density\n(people per km2)")
maphist(v = mtq$POPDENS)
}
\seealso{
\link{getBreaks}, \link{choroLayer}, \link{legendChoro}.
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
