% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{nuts0.spdf}
\alias{nuts0.spdf}
\title{Nuts regions (level 0)}
\format{This basemap is a SpatialPolygonsDataFrame.}
\source{
UMS RIATE
}
\description{
Delineations of EU administrative units (level 0, 2006 version)
}
\details{
This SpatialPolygonsDataFrame can be used with the dataframe
nuts0.df
The projection used is Lambert azimuthal equal-area projection (EPSG:3035).
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique nuts id (character)}

\item{\code{name}}{Official name of the administrative unit}
}}

