% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{nuts2.spdf}
\alias{nuts2.spdf}
\title{Nuts2 Regions}
\format{SpatialPolygonsDataFrame.}
\source{
UMS RIATE
}
\description{
Delineations of EU administrative units (level 2, 2006 version).
}
\details{
This SpatialPolygonsDataFrame can be used with the nuts2.df data.frame
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique nuts id (character)}

\item{\code{name}}{Official name of the administrative unit}
}}

