% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{nuts3.df}
\alias{nuts3.df}
\title{Nuts3 dataset}
\format{This data.frame has 1482 rows and 7 columns.}
\source{
UMS RIATE
}
\description{
This dataset contains some socio-economic data
}
\details{
This DataFrame can be used with the SpatialPolygonsDataFrame
nuts3.spdf
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique nuts id (character)}

\item{\code{birth_2008}}{Number of birth in 2008 (live birth) (numeric)}

\item{\code{death_2008}}{Number of death in 2008 (death) (numeric)}

\item{\code{gdppps1999}}{Gross domestic product (Purchasing Power Standards) in 1999
(million euros) (numeric)}

\item{\code{gdppps2008}}{Gross domestic product (Purchasing Power Standards) in 2008
(million euros) (numeric)}

\item{\code{pop1999}}{Total population in 1999 (inhabitants) (numeric)}

\item{\code{pop2008}}{Total population in 2008 (inhabitants) (numeric)}
}}

