% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmLayer.R
\name{osmLayer}
\alias{osmLayer}
\title{Plot an OpenStreetMap Layer}
\usage{
osmLayer(x, add = FALSE)
}
\arguments{
\item{x}{OpenStreetMap object}

\item{add}{Whether to add the layer to an existing map (TRUE) or not (FALSE)}
}
\description{
Plot an OpenStreetMap Layer
}
\details{
This function is a wrapper for plot.OpenStreetMap
}
\examples{
data("TNdeleg")
}

