% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkLayer.R
\name{propLinkChoroLayer}
\alias{propLinkChoroLayer}
\title{Layer of Proportional Links Colored with a Variable}
\usage{
propLinkChoroLayer(spdf, df, spdfid = NULL, dfidi = NULL, dfidj = NULL,
  var1 = NULL, var2 = NULL, distr = NULL, col = carto.pal("kaki.pal", n1
  = 6), nbclass = 6, method = "quantile", maxlwd = 20)
}
\arguments{
\item{spdf}{Spatial*DataFrame}

\item{df}{DataFrame with Ids and Labels}

\item{spdfid}{Ids of the obj Spatial*DataFrame}

\item{dfidi}{Ids of origins}

\item{dfidj}{Ids of destinations}

\item{var1}{name of the variable corresponding to lwd of the links}

\item{var2}{name of the variable corresponding to discretize}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null,
the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal",
"quantile", "jenks","q6","geom")}

\item{maxlwd}{max lwd of the links}
}
\value{
a plot
}
\description{
Plot a layer of links colors from a discretization
}

