% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkLayer.R
\name{propLinkLayer}
\alias{propLinkLayer}
\title{Proportional Links Layer}
\usage{
propLinkLayer(sldf, df, sldfid = NULL, sldfids, sldfide, dfid = NULL, dfids,
  dfide, var, maxlwd = 40, col, add = T)
}
\arguments{
\item{sldf}{SpatialLinesDataFrame; a link layer.}

\item{df}{DataFrame with identifiers and a variable.}

\item{sldfid}{Unique identifier in sldf (sldfids, sldfide, dfids and dfide are not used).}

\item{sldfids}{Identifier of starting points in sldf (sldfid and dfid are not used).}

\item{sldfide}{Identifier of ending points in sldf (sldfid and dfid are not used).}

\item{dfid}{Unique identifier in df (sldfids, sldfide, dfids and dfide are not used).}

\item{dfids}{Identifier of starting points in df (sldfid and dfid are not used).}

\item{dfide}{Identifier of ending points in df (sldfid and dfid are not used).}

\item{var}{Name of the variable used to plot the links}

\item{maxlwd}{Maximum size of the links.}

\item{col}{Color of the links.}

\item{add}{Add}
}
\description{
Plot a layer of proportionnal links
}
\details{
Unlike many cartography functions, identifiers fields are mandatory.
}
\examples{
data("nuts2006")
data_links <- read.csv("/mnt/data/depot/cartography/inst/extdata/data_links.csv",
                       stringsAsFactors=FALSE)
x <- merge(data_links, data_links, by = "project")
x <- unique(x)
x <- x[x$nuts2.x!=x$nuts2.y,]
x$cpt <- 1
head(x, 50)
xx <- aggregate(x = data.frame(x[,"cpt"], stringsAsFactors = F),
                by = list(x$nuts2.x,x$nuts2.y ), FUN = sum)
names(xx) <- c("i", "j", "fij")
x <- unique(x[,2:3])
ll <- getLinkLayer(spdf = nuts2.spdf, spdf2 = nuts2.spdf, df = x)
xxx <- xx[xx$fij>3,]
propLinkLayer(sldf = ll, df = xxx,
              sldfids = "nuts2.x", sldfide = "nuts2.y",
              dfids = "i", dfide = "j",
              var = "fij", maxlwd = 10, col = "#92000050", add = FALSE)
plot(nuts2.spdf, add = TRUE)
title("ESPON Collab")
}

