% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propSymbolsChoroLayer.R
\name{propSymbolsChoroLayer}
\alias{propSymbolsChoroLayer}
\title{Proportional and Choropleth Symbols Layer}
\usage{
propSymbolsChoroLayer(spdf, df, spdfid = NULL, dfid = NULL, var, k = 0.02,
  fixmax = NULL, symbols = "circle", border = "grey20", lwd = 1, var2,
  breaks = NULL, method = "quantile", nclass = NULL, col = NULL,
  legend.title.cex = 0.8, legend.values.cex = 0.6,
  legend.var.pos = "right", legend.var.title.txt = var,
  legend.var.values.rnd = 0, legend.var.style = "c",
  legend.var.frame = FALSE, legend.var2.pos = "topright",
  legend.var2.title.txt = var2, legend.var2.values.rnd = 0,
  legend.var2.nodata = "no data", legend.var2.frame = FALSE, add = TRUE)
}
\arguments{
\item{spdf}{SpatialPointsDataFrame or SpatialPolygonsDataFrame; if spdf
is a SpatialPolygonsDataFrame symbols are plotted on centroids.}

\item{df}{a data frame that contains the values to plot.}

\item{spdfid}{identifier field in spdf, default to the first column
of the spdf data frame. (optional)}

\item{dfid}{identifier field in df, default to the first column
of df. (optional)}

\item{var}{name of the numeric field in df to plot the symbols sizes.}

\item{k}{share of the map occupied by the biggest symbol (see \link{propSymbolsLayer} Details).}

\item{fixmax}{value of the biggest symbol (see \link{propSymbolsLayer} Details).}

\item{symbols}{type of symbols, one of "circle", "square" or "bar".}

\item{border}{color of symbols borders.}

\item{lwd}{width of symbols borders.}

\item{var2}{name of the numeric field in df to plot the colors.}

\item{breaks}{break points in sorted order to indicate the intervals for assigning the colors.
Note that if there are nlevel colors (classes) there should be (nlevel+1)
breakpoints (see \link{choroLayer} Details).}

\item{method}{a discretization method; one of "sd", "equal",
"quantile", "jenks", "q6" or "geom"  (see \link{choroLayer} Details).}

\item{nclass}{a targeted number of classes. If null, the number of class is
automatically defined (see \link{choroLayer} Details).}

\item{col}{a vector of colors. Note that if breaks is specified there must be one less
colors specified than the number of break.}

\item{legend.title.cex}{size of the legend title.}

\item{legend.values.cex}{size of the values in the legend.}

\item{legend.var.pos}{position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright". If
legend.var.pos is "n" then the legend is not plotted.}

\item{legend.var.title.txt}{title of the legend (proportionnal symbols).}

\item{legend.var.values.rnd}{number of decimal places of the values in
the legend.}

\item{legend.var.style}{either "c" or "e". The legend has two display
styles.}

\item{legend.var.frame}{whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{legend.var2.pos}{position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright". If
legend.var2.pos is "n" then the legend is not plotted.}

\item{legend.var2.title.txt}{title of the legend (colors).}

\item{legend.var2.values.rnd}{number of decimal places of the values in
the legend.}

\item{legend.var2.nodata}{text for "no data" values}

\item{legend.var2.frame}{whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot a proportional symbols layer with color based on a
quantitative data discretization.
}
\examples{
data("nuts2006")
## Exemple 1
# Countries plot
plot(nuts0.spdf, col = "grey60",border = "grey20", add=FALSE)
# Population and growth rate
nuts0.df$cagr <- (((nuts0.df$pop2008 / nuts0.df$pop1999)^(1/9)) - 1) * 100
propSymbolsChoroLayer(spdf = nuts0.spdf, df = nuts0.df,var = "pop2008", var2 = "cagr",
                      legend.var.values.rnd = -3, legend.var2.values.rnd = 2)
}

