\name{spatial-package}
\alias{spatial-package}
\alias{spatial}
\docType{package}
\title{
\packageTitle{spatial}
}
\description{
\packageDescription{spatial}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spatial}
\packageIndices{spatial}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{spatial}

Maintainer: \packageMaintainer{spatial}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
