% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carre.R
\name{squared}
\alias{squared}
\title{Fonction pour elever une valeur au carre}
\usage{
squared(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a numeric
}
\description{
Fonction pour elever une valeur au carre
}
\examples{
squared(2) # 4
}
