% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_arrow.R
\name{tc_arrow}
\alias{tc_arrow}
\title{Plot a north arrow}
\usage{
tc_arrow(pos = "topleft", col, adjust)
}
\arguments{
\item{pos}{position. It can be one of 'topleft', 'top', 'topright', 'right', 'bottomright', 'bottom', 'bottomleft', 'left' or a vector of two coordinates in map units (c(x, y))}

\item{col}{arrow color}

\item{adjust}{object of class \code{sf} or \code{sfc} used to adjust the 
arrow to the real north}
}
\description{
Plot a north arrow.
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq)
tc_arrow(pos = "topright")
}
