% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_credits.R
\name{tc_credits}
\alias{tc_credits}
\title{Plot credits}
\usage{
tc_credits(txt, pos = "bottomleft", col, cex = 0.6, font = 3)
}
\arguments{
\item{txt}{text of the credits, use '\\n' to add line breaks}

\item{pos}{position, one of 'bottomleft', 'bottomright' or 'rightbottom'}

\item{col}{color}

\item{cex}{cex of the credits}

\item{font}{font of the credits}
}
\description{
Plot credits (sources, author, year...).
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq)
tc_credits(txt = "Author\nSources - Year")
}
