% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_borders.R
\name{tc_get_borders}
\alias{tc_get_borders}
\title{Get borders from polygons}
\usage{
tc_get_borders(x, id, outer = FALSE, res = NULL, width = NULL)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{id}{name of the identifier variable in x, default to the first column}

\item{outer}{if TRUE outer borders (non-contiguous polygons borders like
maritime borders) are computed.}

\item{res}{resolution of the grid used to compute outer borders (in x units).
A high resolution will give more detailed borders. (optional)}

\item{width}{maximum distance between used to compute outer borders (in x units).
A higher width will build borders between units that are farther apart. (optional)}
}
\value{
An sf object (MULTILINESTRING) of borders is returned. This object
has three id variables: id, id1 and id2.
id1 and id2 are ids of units that neighbour a border; id is the concatenation
of id1 and id2 (with "_" as separator).
}
\description{
Extract borders between polygons.
}
\examples{
mtq <- tc_import_mtq()
mtq_borders <- tc_get_borders(x = mtq)
plot(mtq_borders)
}
