% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_grid.R
\name{tc_get_grid}
\alias{tc_get_grid}
\title{Get a regular grid}
\usage{
tc_get_grid(x, var, cellsize, type = "regular")
}
\arguments{
\item{x}{an sf object of polygons}

\item{var}{name of the numeric variable(s) in x to adapt to the grid (a vector)}

\item{cellsize}{targeted area of the cell, in map units}

\item{type}{shape of the cell, "regular" for squares, "hexagonal" for hexagons}
}
\value{
A grid is returned as an sf object
}
\description{
Build a regular grid based on an sf object of polygons.
}
\examples{
mtq <- tc_import_mtq()
mtq_grid <- tc_get_grid(x = mtq, var = "POP", cellsize = 5000 * 5000)
mtq_grid$DENS <- mtq_grid$POP / mtq_grid$gridarea * 1000 * 1000
tc_map_c(mtq_grid, "DENS", pal = "Reds 2")
}
