% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_links.R
\name{tc_get_links}
\alias{tc_get_links}
\title{Get a link layer from a data.frame of links.}
\usage{
tc_get_links(x, df, x_id, df_id)
}
\arguments{
\item{x}{an sf object, a simple feature collection.}

\item{df}{a data.frame that contains identifiers of starting and ending points.}

\item{x_id}{name of the identifier variable in x, default to the first column (optional)}

\item{df_id}{names of the identifier variables in df, character vector of length 2, default to 
the two first columns. (optional)}
}
\value{
An sf object is returned, it is composed of df and the sfc (LINESTRING) of links.
}
\description{
Create a link layer from a data.frame of links and an sf object.
}
\examples{
mtq <- tc_import_mtq()
mob <- read.csv(system.file("csv/mob.csv", package="cartography"))
# Select links from Fort-de-France (97209))
mob_97209 <- mob[mob$i == 97209, ]
# Create a link layer
mob_links <- tc_get_links(x = mtq, df = mob_97209)
# Plot the links
tc_map(mtq)
tc_map(mob_links, col = "red4", lwd = 2, add = TRUE)
}
