% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_pencil.R
\name{tc_get_pencil}
\alias{tc_get_pencil}
\title{Get a pencil layer}
\usage{
tc_get_pencil(x, size = 100, buffer = 1000, lefthanded = TRUE)
}
\arguments{
\item{x}{an sf object, a simple feature collection (POLYGON or MULTIPOLYGON).}

\item{size}{density of the penciling. Median number of points used to build
the MULTILINESTRING.}

\item{buffer}{buffer around each polygon. This buffer (in map units) is used
to take sample points. A negative value adds a margin between the penciling
and the original polygons borders}

\item{lefthanded}{if TRUE the penciling is done left-handed style.}
}
\value{
A MULTILINESTRING sf object is returned.
}
\description{
Create a pencil layer. This function transforms a POLYGON or
MULTIPOLYGON sf object into a MULTILINESTRING one.
}
\examples{
mtq <- tc_import_mtq()
mtq_pencil <- tc_get_pencil(x = mtq, size = 100, buffer = 1000,
                            lefthanded = TRUE)
tc_map(mtq_pencil)
}
