% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_init.R
\name{tc_init}
\alias{tc_init}
\title{Plot an invisible layer}
\usage{
tc_init(
  x,
  expandBB = rep(0, 4),
  bg,
  shadow = FALSE,
  shadow_col = "grey50",
  shadow_cex = 1,
  theme,
  export,
  filename,
  width,
  height,
  res = 72
)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}}

\item{expandBB}{fractional values to expand the bounding box with, in each
direction (bottom, left, top, right)}

\item{bg}{background color}

\item{shadow}{add the shadow of x}

\item{shadow_col}{color of the shadow}

\item{shadow_cex}{shadow extent (1 means 1 percent of the shortest side of
the bounding box)}

\item{theme}{apply a theme from \code{tc_theme}}

\item{export}{if set to "png" or "svg" a png or svg plot device is opened}

\item{filename}{path to the exported file}

\item{width}{width of the figure (pixels for png, inches for svg)}

\item{height}{height of the figure (pixels for png, inches for svg)}

\item{res}{resolution (for png)}
}
\description{
Plot an invisible layer with the extent of a spatial object.
Export a map in png or svg format. It uses a reference
geographic layer to find a correct ratio for the export.
}
\examples{
mtq <- tc_import_mtq()
target <- mtq[30, ]
tc_init(target, bg = "lightblue")
tc_map(mtq, add = TRUE)
}
