% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_layout.R
\name{tc_layout}
\alias{tc_layout}
\title{Plot title, credits, scale and arrow}
\usage{
tc_layout(
  title = "Title of the map, year",
  credits = "Authors & Sources",
  scale = TRUE,
  arrow = TRUE,
  frame = FALSE
)
}
\arguments{
\item{title}{title of the map}

\item{credits}{credits}

\item{scale}{display a scale bar}

\item{arrow}{display an arrow}

\item{frame}{display a frame}
}
\description{
Plot a layout layer (title, credits, scalebar, north arrow).
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq) 
tc_layout()
}
