% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_leg_gl.R
\name{tc_leg_gl}
\alias{tc_leg_gl}
\title{Plot a legend for a graduated lines map}
\usage{
tc_leg_gl(
  pos = "topleft",
  val,
  col = "tomato4",
  lwd,
  title = "Title of the Legend",
  title_cex = 0.8,
  val_cex = 0.6,
  val_rnd = 2,
  frame = FALSE,
  bg,
  fg,
  cex = 1
)
}
\arguments{
\item{pos}{position of the legend, one of "topleft", "top",
"topright", "right", "bottomright", "bottom", "bottomleft",
"left" or a vector of two coordinates in map units
(c(x, y))}

\item{val}{break labels}

\item{col}{lines color}

\item{lwd}{lines widths}

\item{title}{title of the legend}

\item{title_cex}{size of the legend title}

\item{val_cex}{size of the values in the legend}

\item{val_rnd}{number of decimal places of the values in
the legend.}

\item{frame}{whether to add a frame to the legend (TRUE) or not (FALSE)}

\item{bg}{background of the legend}

\item{fg}{foreground of the legend}

\item{cex}{size of the legend; 2 means two times bigger}
}
\description{
This function plots a legend for graduated lines maps.
}
\examples{
plot.new()
plot.window(xlim = c(0, 1), ylim = c(0, 1), asp = 1)
tc_leg_gl(lwd =  c(0.2,2,4,5,10), val =c(1,2,3,4,10.2,15.2))
}
