% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_legacy_mode.R
\name{tc_legacy_mode}
\alias{tc_legacy_mode}
\title{Turn on or off warnings from deprecated functions}
\usage{
tc_legacy_mode(mode = "on")
}
\arguments{
\item{mode}{"on" hide deprecation warnings, "off" to show deprecation warnings.}
}
\description{
Use this function to turn off warnings from deprecated function (see Note).
The best way to avoid these warnings is to use the suggested replacement
function.
}
\note{
Be aware that this function may affect other package output. This is 
a wrapper around \code{options(lifecycle_verbosity="quiet")} when 
\code{mode="on"} and \code{options(lifecycle_verbosity="default")} when 
\code{mode="off"}.
}
\examples{
tc_legacy_mode("on")
tc_legacy_mode("off")
}
