% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map.R
\name{tc_map}
\alias{tc_map}
\title{Plot an sf object}
\usage{
tc_map(x, col = "grey80", border = "grey20", lwd = 0.7, bg, add = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{sf} or \code{sfc}}

\item{col}{color}

\item{border}{border color}

\item{lwd}{border width}

\item{bg}{background color}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}

\item{...}{further parameters from \link{plot} for sfc objects}
}
\description{
Plot an sf object. This is mostly a wrapper around
\code{plot(st_geometry(x), ...)}.
}
\examples{
library(sf)
mtq <- tc_import_mtq()
tc_map(mtq)
tc_map(mtq, bg = "green", col = "blue")
}
