% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_dd.R
\name{tc_map_dd}
\alias{tc_map_dd}
\title{Plot a dot dentity map}
\usage{
tc_map_dd(
  x,
  var,
  n,
  pch = 1,
  cex = 0.15,
  type = "random",
  col = "black",
  leg_pos = "topright",
  leg_title,
  leg_cex = 0.6,
  leg_frame = TRUE,
  add = TRUE
)
}
\arguments{
\item{x}{an sf object}

\item{var}{name of the numeric variable to plot.}

\item{n}{one dot on the map represents n (in var units).}

\item{pch}{symbol to use: \link{points}.}

\item{cex}{size of the symbols}

\item{type}{points allocation method: "random" or "regular" (see Details).}

\item{col}{color of the points.}

\item{leg_pos}{"topright", "left", "right", "bottomleft", "bottom", "bottomright". If 
legend.pos is "n" then the legend is not plotted.}

\item{leg_title}{text in the legend.}

\item{leg_cex}{size of the legend text.}

\item{leg_frame}{whether to add a frame to the legend (TRUE) or 
not (FALSE).}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}
}
\description{
Plot a dot density map
}
\details{
The type parameters is defined within the \link{st_sample} function.
}
\examples{
\dontrun{
mtq <- tc_import_mtq()
tc_init(mtq, shadow = TRUE, theme = "green")
tc_map(mtq, add = TRUE)
tc_map_dd(x = mtq,  var="POP", pch=20, col = "red4",n = 200)
tc_layout(title = "Population Distribution in Martinique, 2015")
}
}
