% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_d.R
\name{tc_map_disc}
\alias{tc_map_disc}
\title{Plot a discontinuities map}
\usage{
tc_map_disc(
  x,
  df,
  df_id,
  var,
  breaks = "quantile",
  nbreaks = 4,
  threshold = 0.75,
  type = "rel",
  sizemin = 1,
  sizemax = 10,
  col = "tomato4",
  leg_pos = "bottomleft",
  leg_title = "legend title",
  leg_title_cex = 0.8,
  leg_val_cex = 0.6,
  leg_val_rnd = 2,
  leg_frame = FALSE,
  add = TRUE
)
}
\arguments{
\item{x}{an sf object, a simple feature collection, as outputted by the \link{tc_get_borders} function.}

\item{df}{a data.frame that contains the values used to compute and plot discontinuities.}

\item{df_id}{name of the identifier variable in df, default to the first column 
of df. (optional)}

\item{var}{name of the numeric variable used to compute and plot discontinuities.}

\item{breaks}{a classification method; one of "sd", "equal", "quantile", "fisher-jenks","
q6", "geom", "arith", "em" or "msd" (see \link{getBreaks}).}

\item{nbreaks}{a targeted number of classes. If null, the number of 
class is automatically defined (see \link{tc_get_breaks}).}

\item{threshold}{share of represented borders, value between 0 
(nothing) and 1 (all the discontinuities).}

\item{type}{type of discontinuity measure, one of "rel" or "abs" (see Details).}

\item{sizemin}{thickness of the smallest line.}

\item{sizemax}{thickness of the biggest line.}

\item{col}{color of the discontinuities lines.}

\item{leg_pos}{position of the legend, one of "topleft", "top", 
"topright", "right", "bottomright", "bottom", "bottomleft", "left" or a 
vector of two coordinates in map units (c(x, y)). If 
legend.pos is "n" then the legend is not plotted.}

\item{leg_title}{title of the legend.}

\item{leg_title_cex}{size of the legend title.}

\item{leg_val_cex}{size of the values in the legend.}

\item{leg_val_rnd}{number of decimal places of the values in 
the legend.}

\item{leg_frame}{whether to add a frame to the legend (TRUE) or 
not (FALSE).}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}
}
\value{
An \link{invisible} sf object (MULTISTRING) with the discontinuity measures is returned.
}
\description{
This function computes and plots spatial discontinuities. The 
discontinuities are plotted over a border layer (see \link{tc_get_borders}.
The line widths reflect the ratio or the difference between values of an indicator 
in two neighbouring units.
}
\details{
The "rel" type of discontinuity is the result of pmax(value unit 1 / value unit 2, value unit 2 / value unit 1).\cr
The "abs" type of discontinuity is the result of pmax(value unit 1 - value unit 2, value unit 2 - value unit 1).
}
\examples{
mtq <- tc_import_mtq()
# Get borders
mtq_borders <- tc_get_borders(x = mtq)
# Median Income
tc_map_c(x = mtq, var = "MED", border = "grey", lwd = 0.5,
         breaks = 'equal', nbreaks = 6, leg_pos = "topleft",
         leg_title = "Median Income\n(in euros)" )
# Discontinuities
tc_map_disc(x = mtq_borders, df = mtq, 
            var = "MED", col = "red4", nbreaks=3,
            breaks="equal", threshold = 0.5, sizemin = 0.5,
            sizemax = 10, type = "abs",leg_val_rnd = 0,
            leg_title = "Discontinuities\n(absolute difference)",
            leg_pos = "bottomleft", add = TRUE)
}
