% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_p.R
\name{tc_map_p}
\alias{tc_map_p}
\title{Plot proportional symbols}
\usage{
tc_map_p(
  x,
  var,
  inches = 0.3,
  val_max,
  symbol = "circle",
  col = "tomato4",
  border,
  lwd = 0.7,
  leg_pos = "topleft",
  leg_title = var,
  leg_title_cex = 0.8,
  leg_val_cex = 0.6,
  leg_val_rnd = 0,
  leg_frame = FALSE,
  add
)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{var}{name of the variable to plot}

\item{inches}{size of the biggest symbol (radius for circles, half width for squares) in inches.}

\item{val_max}{maximum value used for proportional symbols}

\item{symbol}{type of symbols, 'circle' or 'square'}

\item{col}{color}

\item{border}{border color}

\item{lwd}{border width}

\item{leg_pos}{position of the legend, one of 'topleft', 'top','topright', 'right', 'bottomright', 'bottom', 'bottomleft', 'left' or a vector of two coordinates in map units (c(x, y)). If leg_pos is 'n' then the legend is not plotted.}

\item{leg_title}{title of the legend}

\item{leg_title_cex}{size of the legend title}

\item{leg_val_cex}{size of the values in the legend}

\item{leg_val_rnd}{number of decimal places of the values in the legend}

\item{leg_frame}{whether to add a frame to the legend (TRUE) or not (FALSE)}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}
}
\description{
Plot proportional symbols.
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq) 
tc_map_p(mtq, "POP")

tc_map(mtq)
tc_map_p(
  x = mtq, var = "POP", inches = .4, symbol = "circle", val_max = 90000,
  col = "tomato1", border = "blue", lwd = 1,
  leg_pos = "right", leg_title = "Population",
  leg_title_cex = 1, leg_val_cex = .8, leg_val_rnd = 0,
  leg_frame = TRUE, add = TRUE
)
}
