% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_pc.R
\name{tc_map_pc}
\alias{tc_map_pc}
\title{Plot proportional symbols using choropleth coloration}
\usage{
tc_map_pc(
  x,
  var,
  inches = 0.3,
  val_max,
  symbol = "circle",
  pal = "Mint",
  breaks = "quantile",
  nbreaks,
  border,
  lwd = 0.7,
  col_na = "white",
  leg_pos = c("topright", "bottomright"),
  leg_title = var,
  leg_title_cex = c(0.8, 0.8),
  leg_val_cex = c(0.6, 0.6),
  leg_val_rnd = c(0, 2),
  leg_no_data = "No data",
  leg_frame = c(FALSE, FALSE),
  add
)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{var}{name of the variable to plot}

\item{inches}{size of the biggest symbol (radius for circles, half width for squares) in inches.}

\item{val_max}{maximum value used for proportional symbols}

\item{symbol}{type of symbols, 'circle' or 'square'}

\item{pal}{a set of colors or a palette name (from \link{tc_get_pal} or \link{hcl.colors})}

\item{breaks}{either a numeric vector with the actual breaks, or a name of a method a classification method; one of 'fixed', 'sd', 'equal', 'pretty', 'quantile', 'kmeans', 'hclust', 'bclust', 'fisher', 'jenks', 'dpih', 'q6', 'geom', 'arith', 'em' or 'msd' (see Details).}

\item{nbreaks}{number of classes}

\item{border}{border color}

\item{lwd}{border width}

\item{col_na}{color for missing values}

\item{leg_pos}{position of the legend, one of 'topleft', 'top','topright', 'right', 'bottomright', 'bottom', 'bottomleft', 'left' or a vector of two coordinates in map units (c(x, y)). If leg_pos is 'n' then the legend is not plotted.}

\item{leg_title}{title of the legend}

\item{leg_title_cex}{size of the legend title}

\item{leg_val_cex}{size of the values in the legend}

\item{leg_val_rnd}{number of decimal places of the values in the legend}

\item{leg_no_data}{label for missing values}

\item{leg_frame}{whether to add a frame to the legend (TRUE) or not (FALSE)}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}
}
\description{
Plot proportional symbols with colors based on a quantitative
data classification.
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq) 
tc_map_pc(mtq, c("POP", "MED"))

tc_map(mtq)
mtq[6, "MED"] <- NA
tc_map_pc(
  x = mtq, var = c("POP", "MED"), inches = .35, border = "tomato4",
  val_max = 90000, symbol = "circle", col_na = "grey", pal = "Cividis",
  breaks = "equal", nbreaks = 4, lwd = 4,
  leg_pos = c("bottomright", "bottomleft"),
  leg_title = c("Population", "Median Income"),
  leg_title_cex = c(0.8, 1),
  leg_val_cex = c(.7, .9),
  leg_val_rnd = c(0, 0),
  leg_no_data = "No data",
  leg_frame = c(TRUE, TRUE),
  add = TRUE
)
}
