% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_r.R
\name{tc_map_r}
\alias{tc_map_r}
\title{Plot a raster}
\usage{
tc_map_r(x, add = FALSE, ...)
}
\arguments{
\item{x}{a RasterLayer or RasterBrick object}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}

\item{...}{bgalpha, interpolate, or other arguments passed to be passed to
\code{\link[raster:plotRGB]{plotRGB}} or  \code{\link[raster:plotRGB]{plot}}}
}
\description{
Plot a raster object over a map (RasterLayer and RasterBrick).
}
\note{
This function is a wrapper for \code{\link[raster:plotRGB]{plotRGB}} 
and  \code{\link[raster:plotRGB]{plot}}
from the raster package. The accuracy of the final plot depends on the 
quality of the \code{*.png} file, the scale of \code{x} and the resolution 
setup of the graphic device.
}
\examples{
mtq <- tc_import_mtq()
file_path <- system.file("img/LogoMartinique.png", package = "cartography")
logo <- tc_get_png(x = mtq, file = file_path)
tc_map_r(logo)
logo1 <- logo$layer.1
tc_map_r(logo1)
}
\seealso{
\link{tc_get_png}
}
