% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_t.R
\name{tc_map_t}
\alias{tc_map_t}
\title{Plot a typology map}
\usage{
tc_map_t(
  x,
  var,
  pal = "Dynamic",
  val_order,
  border,
  pch = 21,
  cex = 1,
  lwd = 0.7,
  cex_na = 1,
  pch_na = 4,
  col_na = "white",
  leg_pos = tc_get_leg_pos(x),
  leg_title = var,
  leg_title_cex = 0.8,
  leg_val_cex = 0.6,
  leg_no_data = "No data",
  leg_frame = FALSE,
  add
)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{var}{name of the variable to plot}

\item{pal}{a set of colors or a palette name (from \link{hcl.colors})}

\item{val_order}{values order, a character vector that matches var modalities}

\item{border}{border color}

\item{pch}{pch type of pch if x is a POINT layer}

\item{cex}{cex cex of the symbols if x is a POINT layer}

\item{lwd}{border width}

\item{cex_na}{cex for NA values if x is a POINT layer}

\item{pch_na}{pch for NA values if x is a POINT layer}

\item{col_na}{color for missing values}

\item{leg_pos}{position of the legend, one of 'topleft', 'top','topright', 'right', 'bottomright', 'bottom', 'bottomleft', 'left' or a vector of two coordinates in map units (c(x, y)). If leg_pos is 'n' then the legend is not plotted.}

\item{leg_title}{title of the legend}

\item{leg_title_cex}{size of the legend title}

\item{leg_val_cex}{size of the values in the legend}

\item{leg_no_data}{label for missing values}

\item{leg_frame}{whether to add a frame to the legend (TRUE) or not (FALSE)}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}
}
\description{
Plot a typology map.
}
\examples{
mtq <- tc_import_mtq()
tc_map_t(mtq, "STATUS")
mtq[6, "STATUS"] <- NA
tc_map_t(
  x = mtq, var = "STATUS", pal = c("red", "blue", "yellow"), lwd = 1.1,
  val_order = c("Prefecture", "Sub-prefecture", "Simple municipality"),
  col_na = "green", border = "brown",
  leg_pos = "bottomleft",
  leg_title = "Status", leg_title_cex = 1.1,
  leg_val_cex = 1, leg_no_data = "No data",
  leg_frame = TRUE, add = FALSE
)
}
