% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_waffle.R
\name{tc_map_waffle}
\alias{tc_map_waffle}
\title{Plot a waffle map}
\usage{
tc_map_waffle(
  x,
  var,
  cell_value,
  cell_size,
  cell_rnd = "ceiling",
  cell_txt = paste0("1 cell = ", cell_value),
  labels,
  ncols,
  pal,
  border = "white",
  lwd = 0.2,
  leg_pos = "bottomleft",
  leg_title = "legend title",
  leg_title_cex = 0.8,
  leg_val_cex = 0.6,
  leg_frame = FALSE,
  add = TRUE
)
}
\arguments{
\item{x}{an sf object, a simple feature collection.}

\item{var}{names of the numeric variable to plot.}

\item{cell_value}{value of a single cell. Original values are rounded, using
\code{cellrnd} method, to be expressed as multiple of \code{cellvalue}.}

\item{cell_size}{size of single cell, in map units.}

\item{cell_rnd}{rounding method, one of "ceiling", "floor", "round".}

\item{cell_txt}{text that appears under the legend.}

\item{labels}{names that will appear in the legend.}

\item{ncols}{number of columns of the waffles}

\item{pal}{a vector of colors.}

\item{border}{color of the cells borders.}

\item{lwd}{cells borders width.}

\item{leg_pos}{position of the legend, one of "topleft", "top", 
"topright", "right", "bottomright", "bottom", "bottomleft", "left" or a 
vector of two coordinates in map units (c(x, y)). If 
legend.pos is "n" then the legend is not plotted.}

\item{leg_title}{title of the legend.}

\item{leg_title_cex}{size of the legend title.}

\item{leg_val_cex}{size of the values in the legend.}

\item{leg_frame}{whether to add a frame to the legend (TRUE) or 
not (FALSE).}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}
}
\description{
Plot a waffle layer.
}
\examples{
mtq <- tc_import_mtq()
# number of employed persons
mtq$EMP <- mtq$ACT - mtq$CHOM
tc_map(mtq,
       col = "#f2efe9",
       border = "#b38e43",
       lwd = 0.5)
tc_map_waffle(
  x = mtq,
  var = c("EMP", "CHOM"),
  cell_value = 100,
  cell_size = 400,
  cell_rnd = "ceiling",
  cell_txt = "1 cell represents 100 persons",
  labels = c("Employed", "Unemployed"),
  ncols = 4,
  pal = c("tomato1", "lightblue"),
  border = "#f2efe9",
  leg_pos = "topright",
  leg_title_cex = .9,
  leg_title = "Active Population",
  leg_val_cex = 0.7,
  add = TRUE
)
tc_layout(
  title = "Structure of the Active Population",
  scale = FALSE,
  credits = "Sources: Insee and IGN, 2018",
)
}
