% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_map_words.R
\name{tc_map_words}
\alias{tc_map_words}
\title{Plot a wordcloud map}
\usage{
tc_map_words(
  x,
  txt,
  freq,
  max.words = NULL,
  cex.maxmin = c(1, 0.5),
  rot.per = 0.1,
  col = NULL,
  fittopol = FALSE,
  use.rank = FALSE,
  add = FALSE,
  breaks = NULL,
  method = "quantile",
  nclass = NULL
)
}
\arguments{
\item{x}{an sf object, a simple feature collection (POLYGON or MULTIPOLYGON).}

\item{txt}{labels variable.}

\item{freq}{frequencies of \code{txt}.}

\item{max.words}{Maximum number of words to be plotted. least frequent terms dropped}

\item{cex.maxmin}{integer (for same size in all \code{txt}) or vector of length 2 indicating the range of the size of the words.}

\item{rot.per}{proportion words with 90 degree rotation}

\item{col}{color or vector of colors words from least to most frequent}

\item{fittopol}{logical. If true would override \code{rot.per} for some elements of \code{x}}

\item{use.rank}{logical. If true rank of frequencies is used instead of real frequencies.}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE)}

\item{breaks, method, nclass}{additional arguments for adjusting the colors of \code{txt}, see \code{\link{choroLayer}}.}
}
\description{
Plot a word cloud adjusted to an \code{sf} object.
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq,
       col = "white",
       bg = "grey95",
       border = NA)
bks <- tc_get_breaks(x = mtq$POP, nbreaks = 5, breaks = "quantile")
cols <- tc_get_pal(n = 5, palette = "Blues", rev = TRUE)
tc_map_words(
  x = mtq,
  txt = "LIBGEO",
  freq = "POP",
  add = TRUE,
  breaks = bks, 
  col = cols 
)
tc_leg_c(
  title = "Population",
  val = bks,
  pal = cols,
  val_rnd = 0,
  no_data = FALSE
)
}
\references{
Ian Fellows (2018). wordcloud: Word Clouds.\cr\cr
R package version 2.6. \url{https://CRAN.R-project.org/package=wordcloud}
}
\seealso{
\link{choroLayer}, \link{legendChoro}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
