% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_pattern.R
\name{tc_pattern}
\alias{tc_pattern}
\title{Plot hatched polygons}
\usage{
tc_pattern(x, pattern = "dot", density = 1, txt = "a", ...)
}
\arguments{
\item{x}{an sf object, a simple feature collection. It should be either a 
\code{POLYGON} or a \code{MULTIPOLYGON}.}

\item{pattern}{Desired pattern to use for hatching. Possible values are:
\itemize{
  \item Dots: \code{"dot", "text"}
  \item Lines \code{"diamond","grid","hexagon","horizontal", "vertical",
  "zigzag","left2right","right2left","circle"}
}}

\item{density}{of the grid. By default the function uses a grid with a 
minimum of 10 cells on the shortest dimension of the bounding box. 
Additionally, it is possible to pass a 
\code{\link[sf:st_make_grid]{cellsize}} value that would feed the 
\code{\link[sf:st_make_grid]{st_make_grid}} underlying function.}

\item{txt}{for the \code{"text"} pattern, that should be a character.}

\item{...}{Additional graphic parameters (see Details).}
}
\value{
When passing \code{mode='sfc'} an 'sf' object (either MULTLINESTRING
 or MULTIPOINT) is returned.
}
\description{
Plot a hatched layer with several different patterns. Suitable 
for b/w print maps.
}
\details{
Possible values are:
\tabular{lccccccc}{
\bold{pattern}        \tab \bold{add} \tab \bold{col} \tab \bold{bg} \tab \bold{cex} \tab \bold{pch} \tab \bold{lwd} \tab \bold{lty}\cr
\bold{"dot"}          \tab x   \tab x   \tab x  \tab x   \tab x   \tab     \tab    \cr
\bold{"text"}         \tab x   \tab x   \tab    \tab x   \tab     \tab     \tab    \cr
\bold{Lines patterns} \tab x   \tab x   \tab    \tab     \tab     \tab x   \tab x  
}
}
\examples{
library(sf)
mtq <- tc_import_mtq()
tc_pattern(mtq, "dot")
title("dot")
plot(st_geometry(mtq), border = NA, col="grey80")
tc_pattern(mtq, "text", txt = "Y", add=TRUE)
title("text")
tc_pattern(mtq, "diamond", density = 0.5)
plot(st_union(st_geometry(mtq)), add = TRUE)
title("diamond")
tc_pattern(mtq, "grid", lwd = 1.5)
title("grid")
tc_pattern(mtq, "hexagon", col = "blue")
title("hexagon")
tc_pattern(mtq, "horizontal", lty = 5)
title("horizontal")
tc_pattern(mtq, "vertical")
title("vertical")
tc_pattern(mtq, "left2right")
title("left2right")
tc_pattern(mtq, "right2left")
title("right2left")
tc_pattern(mtq, "zigzag",cellsize=5000)
title("zigzag")
tc_pattern(mtq, "circle")
title("circle")
}
\seealso{
\link{tc_leg_pattern}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
