% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_theme.R
\name{tc_theme}
\alias{tc_theme}
\title{Set a theme}
\usage{
tc_theme(x)
}
\arguments{
\item{x}{name of a map theme. One of "default", "brutal", "ink",
"dark", "agolalight", "candy", "darkula", "iceberg", "green", "nevermind",
"jsk", "barcelona".
It is also possible to set a custom theme (see Examples).
The parameters set by this function are the figure margins, background and
foreground colors and some \link{tc_title} options.
Use \code{tc_theme('default')} to reset theme settings.}
}
\value{
The (invisible) list of parameters of the current theme.
}
\description{
This function set a color theme.
}
\examples{
mtq <- tc_import_mtq()
tc_theme("green")
tc_map(mtq)
tc_title()

custom <- list(
  name = "custom",
  bg = "green",
  fg = "red",
  mar = c(2, 2, 2, 2),
  tab = TRUE,
  pos = "center",
  inner = TRUE,
  line = 2,
  cex = 1.5,
  font = 3
)
tc_theme(custom)
tc_map(mtq)
tc_title()
}
