% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_title.R
\name{tc_title}
\alias{tc_title}
\title{Plot a title}
\usage{
tc_title(txt = "Map Title, Year", pos, tab, bg, fg, cex, line, font, inner, x)
}
\arguments{
\item{txt}{title text}

\item{pos}{position, one of 'left', 'center', 'right'}

\item{tab}{if TRUE the title is displayed as a 'tab'}

\item{bg}{background of the title}

\item{fg}{foreground of the title}

\item{cex}{cex of the title}

\item{line}{number of lines used for the title}

\item{font}{font of the title}

\item{inner}{if TRUE the title is displayed inside the plot area.}

\item{x}{object of class \code{sf} or \code{sfc}}
}
\description{
Plot a title
}
\examples{
mtq <- tc_import_mtq()
tc_map(mtq)
tc_title()
}
