% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmLayer.R
\name{tilesLayer}
\alias{tilesLayer}
\title{Plot Tiles from Open Map Servers}
\usage{
tilesLayer(x, add = FALSE)
}
\arguments{
\item{x}{an OpenStreetMap object; the \link{getTiles} function outputs these
objects.}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot tiles from open map servers.
}
\note{
This function is a wrapper for plot.OpenStreetMap from the
OpenStreetMap package.
}
\examples{
\dontrun{
data("nuts2006")
# Download the tiles
StamWatCol <- getTiles(spdf = nuts0.spdf, type = "stamen-watercolor")
#Display the tiles Layer
tilesLayer(x = StamWatCol)
#Display countries boundaries
plot(nuts0.spdf, add = TRUE)
}
}

