% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesLayer.R
\name{tilesLayer2}
\alias{tilesLayer2}
\title{Plot Tiles from Open Map Servers 2}
\usage{
tilesLayer2(x, add = FALSE)
}
\arguments{
\item{x}{a RasterBrick object; the \link{getTiles2} function outputs these 
objects.}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}
}
\description{
Plot tiles from open map servers 2.
}
\note{
This function is a wrapper for plotRGB from the 
raster package.
}
\seealso{
\link{getTiles}
}

