% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{TNdeleg}
\alias{TNdeleg}
\title{Tunisian Delegations Dataset}
\format{This data.frame has 263 rows and 13 columns.}
\source{
Institut National de la Statistique (INS). General census of
population and housing 2004 (last update 18-06-2013). Data gathered on
October 2014.
}
\usage{
data(TNdeleg)
}
\description{
This dataset contains some socio-economic data on tunisian
delegations in 2004 (population, active population, unemployed population,
housing, households).
}
\details{
This data.frame can be joined to the \code{\link{TNdeleg.spdf}}
SpatialPolygonsDataframe. The \code{del_code} field from TNDeleg matches
the \code{del_code} field from TNdeleg.spdf.
}
\section{Fields}{

\describe{
\item{\code{del_code}}{delegation code (character)}

\item{\code{del_name}}{delegation name (character)}

\item{\code{pop_m}}{male population (numeric)}

\item{\code{pop_f}}{female population (numeric)}

\item{\code{pop_t}}{total population (numeric)}

\item{\code{act_m}}{active male population of 15 years old and over (numeric)}

\item{\code{act_f}}{active female population of 15 years old and over (numeric)}

\item{\code{act_t}}{total active population of 15 years old and over (numeric)}

\item{\code{unem_m}}{unemployed male population of 15 years old and over (numeric)}

\item{\code{unem_f}}{unemployed female population of 15 years old and over (numeric)}

\item{\code{unem_t}}{unemployed total population of 15 years old and over (numeric)}

\item{\code{house}}{number of household (numeric)}

\item{\code{housing}}{number of housing (numeric)}
}}
\seealso{
\code{\link{TNdeleg.spdf}}
}

