% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{TNdeleg.spdf}
\alias{TNdeleg.spdf}
\title{Tunisian Delegations Basemap}
\format{This basemap is a SpatialPolygonsDataFrame.}
\source{
Unknown
}
\usage{
data(TNdeleg)
}
\description{
Basemap of the tunisian delegations
}
\details{
This SpatialPolygonsDataFrame can be joined to the
\code{\link{TNdeleg}} data.frame.
The \code{del_code} field from TNDeleg matches the \code{del_code} field
from TNdeleg.spdf.
The projection used is Lambert azimuthal equal-area projection (EPSG:3035).
}
\section{Fields}{

\describe{
\item{\code{del_code}}{delegation code (character)}
}}
\seealso{
\code{\link{TNdeleg}}
}

