% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twincities.df}
\alias{twincities.df}
\title{Twin Cities Dataset}
\source{
Adam Ploszaj - Centre for European Regional and Local Studies EUROREG, 
University of Warsaw, Poland. 
Primary source: Wikipedia, 2011.
}
\description{
This dataset contains the number of international twinning 
agreements betwwen cities. Agreements are aggregated at nuts2 level.
}
\details{
This data frame can be used with the SpatialPolygonsDataFrame 
nuts2.spdf
}
\section{Fields}{

\describe{
\item{\code{i}}{nuts2 identifier}

\item{\code{j}}{nuts2 identifier}

\item{\code{fij}}{number of agreements}
}}

