% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffleLayer.R
\name{waffleLayer}
\alias{waffleLayer}
\title{Waffle Layer}
\usage{
waffleLayer(
  x,
  var,
  value,
  labels,
  ncols = 5,
  cellsize,
  col,
  border = "white",
  lwd = 0.2,
  legend.pos = "bottomleft",
  legend.title.txt = var,
  legend.title.cex = 0.8,
  legend.values.cex = 0.6,
  legend.frame = FALSE,
  add = FALSE
)
}
\arguments{
\item{x}{an sf object, a simple feature collection.}

\item{var}{names of the numeric variable to plot.}

\item{value}{value of a single cell (variables will be rounded to be 
expressed as multiple of value). See Details}

\item{labels}{names that will appear in the legend.}

\item{ncols}{number of columns of the waffles}

\item{cellsize}{size of single cell, in map units.}

\item{col}{a vector of colors.}

\item{border}{color of the polygons borders.}

\item{lwd}{borders width.}

\item{legend.pos}{position of the legend, one of "topleft", "top", 
"topright", "right", "bottomright", "bottom", "bottomleft", "left" or a 
vector of two coordinates in map units (c(x, y)). If 
legend.pos is "n" then the legend is not plotted.}

\item{legend.title.txt}{title of the legend.}

\item{legend.title.cex}{size of the legend title.}

\item{legend.values.cex}{size of the values in the legend.}

\item{legend.frame}{whether to add a frame to the legend (TRUE) or 
not (FALSE).}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}
}
\description{
Plot a waffle layer.
}
\examples{
library(sf)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package="cartography"))
mtq$EMP <- mtq$ACT - mtq$CHOM
mtq$INACT <- mtq$POP - mtq$ACT

par(mar=c(0,0,0,0))

plot(st_geometry(mtq), col="#f2efe9", border="#b38e43", bg = "#aad3df",
     lwd = 0.5)
waffleLayer(x = mtq, var = rev(c("CHOM","EMP", "INACT")), 
            value = 200, cellsize = 300, ncols = 10)
}
