% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wordlayout}
\alias{wordlayout}
\title{Title}
\usage{
wordlayout(x, y, words, cex = 1, rotate90 = FALSE, xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf), tstep = 0.1, rstep = 0.1, ...)
}
\arguments{
\item{x}{t}

\item{y}{t}

\item{words}{t}

\item{cex}{t}

\item{rotate90}{t}

\item{xlim}{t}

\item{ylim}{t}

\item{tstep}{t}

\item{rstep}{t}

\item{...}{t}
}
\value{
sdsdss
}
\description{
Title
}
