% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world.spdf}
\alias{world.spdf}
\title{World Background}
\format{
SpatialPolygonsDataFrame.
}
\source{
UMS RIATE - \url{http://riate.cnrs.fr/}
}
\description{
World background.
}
\keyword{internal}
