% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\name{carto.pal}
\alias{carto.pal}
\title{carto.pal}
\usage{
carto.pal(pal1, n1, pal2 = NULL, n2 = NULL, middle = FALSE,
  alphaeffect = FALSE)
}
\arguments{
\item{pal1}{name of the color gradiant}

\item{n1}{number of colors (up to 20)}

\item{pal2}{name of a second color gradiant (optionnal)}

\item{n2}{number of colors for the second color gradiant (up to 20) (optionnal)}

\item{middle}{boolean. if true, add a neutral color in the middle (optionnal)}

\item{alphaeffect}{boolean. If true, enhance contrasts by adding opacity (optionnal)}
}
\value{
vector
}
\description{
Build a color palette
}
\details{
This function call and combine colors from the colors object with several paraemeters
}
\examples{
# Simple gradient in blue
carto.pal("blue.pal",20)

# double gradient blue / red
carto.pal(pal1="blue.pal",n1=10, pal2="red.pal",n2=10)

# Adding a neutral color
carto.pal(pal1="blue.pal",n1=10,pal2="red.pal",n2=10,middle=TRUE)

# Enhancing contrasts with alphaeffect
carto.pal(pal1="blue.pal",n1=10,pal2="red.pal",n2=10,middle=TRUE,alphaeffect=TRUE)

# The double gradient can be asymetric
carto.pal(pal1="blue.pal",n1=5,pal2="red.pal",n2=15,middle=TRUE,alphaeffect=TRUE)

# Build and display a palette
mypal <- carto.pal(pal1="blue.pal",n1=5,pal2="red.pal",n2=15,middle=TRUE,alphaeffect=TRUE)
k<-length(mypal)
image(1:k, 1, as.matrix(1:k), col =mypal, xlab = paste(k," classes",sep=""),
ylab = "", xaxt = "n", yaxt = "n",bty = "n")
}

