% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cartography.colors}
\alias{cartography.colors}
\title{Color Palettes}
\source{
UMS RIATE
}
\description{
List of color gradients adapted to thematic cartography.
}
\seealso{
\link{display.carto.pal}, \link{display.carto.all}, \link{carto.pal}
}

