% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{cartography}
\alias{cartography}
\alias{cartography-package}
\title{Cartography Package}
\description{
Create and integrate maps into your R workflow. Classical
thematic maps and more unconventionnal geographical representation are
available.
}

