% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choroLayer.R
\name{choroLayer}
\alias{choroLayer}
\title{choroLayer}
\usage{
choroLayer(spdf, df, spdfid = NULL, dfid = NULL, var, distr = NULL,
  col = NULL, nbclass = NULL, method = "quantile", add = FALSE)
}
\arguments{
\item{spdf}{spatial dataframe (geometries)}

\item{df}{dataframe (data)}

\item{spdfid}{id of the spdf (if null, the first column is used)}

\item{dfid}{id of the df (if null, the first column is used)}

\item{var}{field used}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null,
the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal",
"quantile", "jenks","q6","geom")}

\item{add}{add = T}
}
\value{
plot
}
\description{
plot a chorpoleth layer
}
\details{
plot a chorpoleth layer
}
\examples{
choroLayer(spdf = TNdeleg.spdf, df = TNdeleg, var =  "housing", nbclass = 20,
method = "equal",add=FALSE)
}

