% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coasts.spdf}
\alias{coasts.spdf}
\title{Coastline of Europe}
\format{
SpatialLinesDataFrame.
}
\source{
UMS RIATE - \url{https://riate.cnrs.fr/}
}
\description{
Coastline of Europe.
}
\keyword{internal}
