% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries.spdf}
\alias{countries.spdf}
\title{Countries in the European Area}
\format{SpatialPolygonsDataFrame.}
\source{
UMS RIATE - \url{http://www.ums-riate.fr/Webriate/?page_id=153}
}
\description{
Countries in the european area.
}

