% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discLayer.R
\name{discLayer}
\alias{discLayer}
\title{Spatial discontinuities}
\usage{
discLayer(spdf, df, spdfid1 = NULL, spdfid2 = NULL, dfid = NULL, var,
  distr = NULL, col = "red", nbclass = NULL, method = "quantile",
  threshold = 0.75, sizemin = 1, sizemax = 10, type = "rel",
  add = FALSE)
}
\arguments{
\item{spdf}{SpatialLinesDataFrame (geometric borders)}

\item{df}{dataframe (data)}

\item{spdfid1}{first id of the border}

\item{spdfid2}{second id of the border}

\item{dfid}{id of the df (if null, the first column is used)}

\item{var}{field used}

\item{distr}{vector of classes}

\item{col}{Color (character)}

\item{nbclass}{number of classes targeted (if null,
the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal",
"quantile", "jenks","q6","geom")}

\item{threshold}{Value between 0 (nothing) and 1 (all the discontinuities). Default = 0.75}

\item{sizemin}{thickness of the smaller line (default = 1)}

\item{sizemax}{thickness of the bigger line (default = 10)}

\item{type}{Mesuer of discontinuity. "rel","abs" (default = "rel")}

\item{add}{add=TRUE}
}
\value{
plot
}
\description{
Compute and plot spatial discontinuities
}
\details{
Compute and plot spatial discontinuities (relative or absolutes)
}
\examples{
data(nuts2006)
nuts0.contig.spdf <-borders(nuts0.spdf)
plot(nuts0.spdf, col="#CCCCCC", lwd=1, border="white")
discLayer(spdf = nuts0.contig.spdf, df = nuts0.df, dfid = "id", spdfid1 = "id1", spdfid2 = "id2",
var = "gdppps2008", col="red", nbclass=5,  method="quantile", threshold = 0.5, sizemin = 1,
sizemax = 10, type = "rel", add=TRUE )
}

