% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.carto.all}
\alias{display.carto.all}
\title{display.carto.all}
\usage{
display.carto.all(nb = 10)
}
\arguments{
\item{nb}{numeric number of colors by palettes (from 1 to 20)}
}
\value{
vector
}
\description{
Display all the color palettes available
}
\details{
This function displays all the palettes available in the package
}
\examples{
display.carto.all(1)
display.carto.all(5)
display.carto.all(8)
display.carto.all(12)
display.carto.all(20)
}

