% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.carto.all}
\alias{display.carto.all}
\title{Display all Cartographic Palettes}
\usage{
display.carto.all(n = 10)
}
\arguments{
\item{n}{number of colors in the gradients (from 1 to 20).}
}
\description{
Display all the available color palettes.
}
\examples{
display.carto.all(1)
display.carto.all(5)
display.carto.all(8)
display.carto.all(12)
display.carto.all(20)
}
\seealso{
\link{carto.pal}, \link{display.carto.pal}, \link{carto.pal.info}
}

