% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.carto.pal}
\alias{display.carto.pal}
\title{display.carto.pal}
\usage{
display.carto.pal(name)
}
\arguments{
\item{name}{character name of the palette available in the package}
}
\value{
plot
}
\description{
Display one color palette
}
\details{
This function displays all the content of one palette
}
\examples{
display.carto.pal("orange.pal")
display.carto.pal("sand.pal")
}

