% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display.carto.pal}
\alias{display.carto.pal}
\title{Display one Cartographic Palette}
\usage{
display.carto.pal(name)
}
\arguments{
\item{name}{name of the palette available in the package (see Details).}
}
\description{
Display one color palette.
}
\details{
Sequential palettes: 
\itemize{
\item{blue.pal}
\item{orange.pal}
\item{red.pal}
\item{brown.pal}
\item{green.pal}
\item{purple.pal}
\item{pink.pal}
\item{wine.pal}
\item{grey.pal}
\item{turquoise.pal}
\item{sand.pal}
\item{taupe.pal}
\item{kaki.pal}
\item{harmo.pal}}
 
Qualitative palettes: 
\itemize{
\item{pastel.pal}
\item{multi.pal}
}
}
\examples{
display.carto.pal("orange.pal")
display.carto.pal("sand.pal")
}
\seealso{
\link{carto.pal}, \link{display.carto.all}, \link{carto.pal.info}
}

