% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frame.spdf}
\alias{frame.spdf}
\title{Frame around Europe}
\format{
SpatialPolygonsDataFrame.
}
\source{
UMS RIATE - \url{http://riate.cnrs.fr/?page_id=153}
}
\description{
Frame around European countries.
}
\keyword{internal}
