% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/borders.R
\name{getBorders}
\alias{getBorders}
\title{Extract SpatialPolygonsDataFrame Borders}
\usage{
getBorders(spdf, spdfid = NULL, tol = 1)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame.}

\item{spdfid}{character; id field in spdf, default to the first column
of the spdf data.frame. (optional)}

\item{tol}{numeric; tolerance to detect contiguity (in map units). You may
not want to change this parameter.}
}
\value{
A SpatialLinesDataFrame of borders is returned. This object has three id fields: id, id1 and id2.
id1 and id2 are ids of units that neighbour a border; id concatenates id1 and id2.
}
\description{
Extract borders between SpatialPolygonsDataFrame units.
}
\examples{
data(nuts2006)
nuts0.contig.spdf <- getBorders(nuts0.spdf)
nuts0.contig.spdf$col <- sample(x = rainbow(length(nuts0.contig.spdf)))
plot(nuts0.spdf, border = NA, col = "grey60")
plot(nuts0.contig.spdf, col = nuts0.contig.spdf$col, lwd = 3, add = TRUE)
}

