% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridLayer.R
\name{getGridData}
\alias{getGridData}
\title{Compute Data for a Grid Layer}
\usage{
getGridData(x, df, dfid = NULL, var)
}
\arguments{
\item{x}{...}

\item{df}{...}

\item{dfid}{...}

\item{var}{...}
}
\description{
Defunct
}
