% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridLayer.R
\name{getGridLayer}
\alias{getGridLayer}
\title{Build A Regular Grid Layer}
\usage{
getGridLayer(spdf, cellsize, spdfid = NULL)
}
\arguments{
\item{spdf}{a SpatialPolygonsDataFrame.}

\item{cellsize}{output cell size, in map units.}

\item{spdfid}{identifier field in spdf, default to the first column
of the spdf data.frame. (optional)}
}
\value{
A list is returned. The list contains spdf: a SpatialPolygonsDataFrame of
a regular grid and df;  a data.frame of surface intersection. df fields are id_cell: ids of the grid;
id_geo: ids of the spdf and area_pct: share of the area of the polygon in the cell
(a value of 55 means that 55\% of the spdf unit is within the cell).
}
\description{
Build a regular grid based on a SpatialPolygonsDataFrame.
Provide also a table of surface intersections.
}
\examples{
data(nuts2006)
mygrid <- getGridLayer(spdf = nuts2.spdf, cellsize = 200000)
plot(mygrid$spdf)
head(mygrid$df)
}

