% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmLayer.R
\name{getTiles}
\alias{getTiles}
\title{Get Tiles from Open Map Servers}
\usage{
getTiles(spdf, type = "osm", zoom = NULL)
}
\arguments{
\item{spdf}{a Spatial*DataFrame with a valid projection attribute.}

\item{type}{the tile server from which to get the map
(see Details).}

\item{zoom}{the zoom level. If null, it is determined automatically
(see Details).}
}
\value{
An OpenStreetMap object is returned.
}
\description{
Get map tiles based on a Spatial*DataFrame extent. Maps can be
fetched from various open map servers.
}
\details{
Available tile servers are described in the openmap function of
the OpenStreetMap package.

Zoom levels are descibed on the OpenStreetMap wiki:
\url{http://wiki.openstreetmap.org/wiki/Zoom_levels}.
}
\note{
This function uses rgdal and OpenStreetMap packages.
}
\examples{
\dontrun{
data("nuts2006")
# Download the tiles
StamWatCol <- getTiles(spdf = nuts0.spdf, type = "stamen-watercolor")
class(StamWatCol)
}
}

