% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/osmLayer.R
\name{getTiles}
\alias{getTiles}
\title{Get Tiles From Open Map Servers}
\usage{
getTiles(spdf, type = "osm", zoom = NULL)
}
\arguments{
\item{spdf}{Spatial*DataFrame; it must have a valid projection attribute.}

\item{type}{character; the tile server from which to get the map
(see Details).}

\item{zoom}{numeric; the zoom level. If null, it is determined automatically
(see Details).}
}
\value{
An OpenStreetMap object is returned.
}
\description{
Get map tiles based on a Spatial*DataFrame extent. Maps can be
fetched from various open map servers.
}
\details{
This function uses rgdal and OpenStreetMap packages.

Available tile servers are described in the \code{openmap} function of
the OpenStreetMap package.

Zoom levels are descibed on the OpenStreetMap wiki:
\url{http://wiki.openstreetmap.org/wiki/Zoom_levels}.
}
\examples{
data("nuts2006")
}

