% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradLinkLayer.R
\name{gradLinkLayer}
\alias{gradLinkLayer}
\title{Proportional Links Layer}
\usage{
gradLinkLayer(sldf, df, sldfid = NULL, sldfids, sldfide, dfid = NULL, dfids,
  dfide, var, breaks, lwd, col, legend.pos = "bottomleft",
  legend.title.txt = var, legend.title.cex = 0.8, legend.values.cex = 0.6,
  legend.frame = FALSE, legend.values.rnd = 0, add = TRUE)
}
\arguments{
\item{sldf}{SpatialLinesDataFrame; a link layer.}

\item{df}{a data. frame with identifiers and a variable.}

\item{sldfid}{unique identifier in sldf (sldfids, sldfide, dfids and dfide are not used).}

\item{sldfids}{identifier of starting points in sldf (sldfid and dfid are not used).}

\item{sldfide}{identifier of ending points in sldf (sldfid and dfid are not used).}

\item{dfid}{unique identifier in df (sldfids, sldfide, dfids and dfide are not used).}

\item{dfids}{identifier of starting points in df (sldfid and dfid are not used).}

\item{dfide}{identifier of ending points in df (sldfid and dfid are not used).}

\item{var}{name of the variable used to plot the links}

\item{breaks}{break values in sorted order to indicate the intervals for assigning the lines widths.}

\item{lwd}{vector of widths.}

\item{col}{color of the links.}

\item{legend.pos}{position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright". If
legend.pos is "n" then the legend is not plotted.}

\item{legend.title.txt}{title of the legend.}

\item{legend.title.cex}{size of the legend title.}

\item{legend.values.cex}{size of the values in the legend.}

\item{legend.frame}{whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{legend.values.rnd}{number of decimal places of the values
displayed in the legend.}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot a layer of proportionnal links
}
\details{
Unlike most of cartography functions, identifiers fields are mandatory.
}
\examples{
data("nuts2006")
# Create a link layer
twincities.spdf <- getLinkLayer(spdf = nuts2.spdf, df = twincities[,1:2])
# plot the links
plot(nuts0.spdf, col = "grey60",border = "grey20")
gradLinkLayer(sldf = twincities.spdf, df = twincities,
              sldfids = "i", sldfide = "j",
              dfids = "i", dfide = "j",legend.pos = "topright",
              var = "fij", breaks = c(2,5,15,20,30), lwd = c(0.1,1,4,10),
              col = "#92000090", add = TRUE)
}

