% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graticule.spdf}
\alias{graticule.spdf}
\title{Graticule around Europe}
\format{SpatialLines.}
\source{
UMS RIATE - \url{http://www.ums-riate.fr/Webriate/?page_id=153}
}
\description{
Graticule around Europe.
}

