% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelLayer.R
\name{labelLayer}
\alias{labelLayer}
\title{labelLayer function.}
\usage{
labelLayer(spdf, df, spdfid = NULL, dfid = NULL, txt, col = "black",
  cex = 0.7, ...)
}
\arguments{
\item{spdf}{Spatial*DataFrame.}

\item{df}{data.frame; \code{df} contains the labels field to plot.}

\item{spdfid}{character; id field in \code{spdf}, default to the first column
of the \code{spdf} data.frame. (optional)}

\item{dfid}{character; id field in \code{df}, default to the first column
of \code{df}. (optional)}

\item{txt}{character; labels field in \code{df}.}

\item{col}{character; labels color.}

\item{cex}{numeric; labels cex.}

\item{...}{further \link{text} arguments, such as \code{pos} or \code{adj}.}
}
\description{
Put labels
}
\examples{
data("nuts2006")

# Display countries boundaries
plot(nuts0.spdf, col = "grey60",border = "grey20")

# Layout plot
layoutLayer(title = "Most Populated Countries of Europe",
            sources = "The 10 most populated countries of Europe",
            author = "Total population 2008, in millions of inhabitants.",
            scale = NULL,
            frame = TRUE,
            col = "black",
            coltitle = "white",
            south = TRUE)

# Selection of the 10 most populated countries of Europe
dflab <- nuts0.df[order(nuts0.df$pop2008, decreasing = TRUE),][1:10,]

# Label creation
dflab$lab <- paste(dflab$id, "\\n", round(dflab$pop2008/1000000,0), "M", sep ="")

# Label plot of the 10 most populated countries
labelLayer(spdf = nuts0.spdf, df = dflab, txt = "lab",
           col = "blue", cex = 0.7, font = 2)
}

