% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/layoutLayer.R
\name{layoutLayer}
\alias{layoutLayer}
\title{layoutLayer function.}
\usage{
layoutLayer(title = "Title of the map, year", sources = "Source(s)",
  author = "Author(s)", col = "black", coltitle = "white", bg = NULL,
  scale = 0, frame = TRUE, north = FALSE, south = FALSE,
  extent = NULL)
}
\arguments{
\item{title}{Title of the map}

\item{sources}{Sources of the map}

\item{author}{Author of the map}

\item{col}{Color of the frame}

\item{coltitle}{Color of the Title}

\item{bg}{Background color}

\item{scale}{Size of the scale in kilometers. If set to NULL, no scale is
displayed, if set to 0 an automatic scale is displayed
(1/10 of the map width)}

\item{frame}{Wheither displaying a frame or not}

\item{north}{Noth arrow}

\item{south}{South arrow}

\item{extent}{extent optional, set the extent to the one of a
Spatial*DataFrame. If not set, plot.new has to be called first.}
}
\description{
layoutLayer function.
}
\details{
Must only be called after a plot
}
\examples{
data("TNdeleg")
}

