% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legends.R
\name{LegendPropLines}
\alias{LegendPropLines}
\title{LegendPropLines}
\usage{
LegendPropLines(pos = "topleft", legTitle = "Title of the legend",
  legTitleCex = 0.8, legValuesCex = 0.6, varvect, sizevect, col = "red",
  frame = FALSE, round = 0)
}
\arguments{
\item{pos}{bottomleft, topleft, topright, bottomright, left, right, top, bottom, middle}

\item{legTitle}{Title of the legend}

\item{legTitleCex}{legTitleCex}

\item{legValuesCex}{legValuesCex}

\item{varvect}{Values}

\item{sizevect}{Sizes of the squares (r)}

\item{col}{color defalul = "red"}

\item{frame}{if TRUE, a frame is drawn}

\item{round}{round}
}
\value{
plot
}
\description{
legend for Prop Lines layers
}
\details{
plot a legend
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
rect(par()$usr[1], par()$usr[3], par()$usr[2], par()$usr[4], border = "black")

LegendPropLines(pos = "topleft", legTitle = "Title\\nxkkxkx",
                legTitleCex = 0.8, legValuesCex = 0.6,
                varvect = nuts1.df$pop2008,
                sizevect = nuts1.df$pop2008/1000000,
                col="red", frame=TRUE, round=2)
}

