% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legends.R
\name{LegendSquaresSymbols}
\alias{LegendSquaresSymbols}
\title{LegendSquaresSymbols}
\usage{
LegendSquaresSymbols(pos = "topleft", legTitle = "Title of the legend",
  legTitleCex = 0.8, legValuesCex = 0.6, varvect, sizevect, breakval,
  col1 = "red", col2 = "blue", frame = FALSE, round = 0, type = "a")
}
\arguments{
\item{pos}{bottomleft, topleft, topright, bottomright, left, right, top, bottom, middle}

\item{legTitle}{Title of the legend}

\item{legTitleCex}{legTitleCex}

\item{legValuesCex}{legValuesCex}

\item{varvect}{Values}

\item{sizevect}{Sizes of the squares (r)}

\item{breakval}{breakval}

\item{col1}{color defalul = "red"}

\item{col2}{color default = "blue"}

\item{frame}{if TRUE, a frame is drawn}

\item{round}{round}

\item{type}{Type of display of the legend. "a" or "b"}
}
\value{
plot
}
\description{
legend for Prop Symbols layers (squares)
}
\details{
plot a legend
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
rect(par()$usr[1], par()$usr[3], par()$usr[2], par()$usr[4], border = "black")
LegendSquaresSymbols(pos = "bottomright", legTitle = "Title of\\nthe legend ",
                     legTitleCex = 0.8, legValuesCex = 0.6,
                     varvect = nuts1.df$pop2008,
                     sizevect = sqrt((abs(nuts1.df$pop2008) * 5000)),
                     breakval=10, col1="red", col2="blue", frame=TRUE, round=0, type ="a")

LegendSquaresSymbols(pos = "topright", legTitle = "Title",
                     legTitleCex = 1.2, legValuesCex = 0.8,
                     varvect = nuts1.df$pop2008,
                     sizevect = sqrt((abs(nuts1.df$pop2008) * 50000)),
                     breakval=10, col1="red", col2="blue", frame=TRUE, round=0, type ="b")

LegendSquaresSymbols(pos = "topleft", legTitle = "Title\\nof\\nthe legend",
                     legTitleCex = 0.8, legValuesCex = 0.6,
                     varvect = nuts1.df$pop2008,
                     sizevect = sqrt((abs(nuts1.df$pop2008) * 500)),
                     breakval=NULL, col1="red", col2="blue", frame=TRUE, round=0, type ="b")
}

