% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legends.R
\name{legendTypo}
\alias{legendTypo}
\title{Legend For Typology Maps}
\usage{
legendTypo(pos = "topleft", title.txt = "Title of the legend",
  title.cex = 0.8, values.cex = 0.6, col, categ, cex = 1, nodata = TRUE,
  nodata.txt = "No data", nodata.col = "white", frame = FALSE,
  symbol = "box")
}
\arguments{
\item{pos}{character; position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{title.txt}{character; title of the legend.}

\item{title.cex}{numeric; size of the legend title.}

\item{values.cex}{numeric; size of the values in the legend.}

\item{col}{character; a vector of colors.}

\item{categ}{vector; vector of categories.}

\item{cex}{character; size of the legend. 2 means two times bigger.}

\item{nodata}{boolean; if TRUE a "no data" box or line is plotted.}

\item{nodata.txt}{character; label for "no data" values.}

\item{nodata.col}{character; color of "no data" values.}

\item{frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{symbol}{character; 'line' or 'box'}
}
\description{
Plot legend for typology maps.
}
\examples{
data("nuts2006")
plot(nuts0.spdf, col = "grey")
box()

jazz <- c("John coltrane","Miles Davis","Maceo Parker",
          "Herbie\\nHancock","Julien Lourau")
legendTypo(pos = "bottomleft", title.txt = "Title of the legend", title.cex = 0.8,
           values.cex = 0.6, categ = jazz, cex = 0.75,
           col = c("red","blue","green","yellow","purple"),
           nodata = TRUE, nodata.txt = "No data\\navailable", frame = TRUE, symbol="box")
legendTypo(pos = "topright", title.txt = "Jazzmen",
           title.cex = 1.5, cex = 1.25,
           values.cex = 1, categ = jazz,
           col = c("red","blue","green","yellow","purple"),
           nodata = FALSE, frame = FALSE, symbol="line")
}

