% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legends.R
\name{LegendTypo}
\alias{LegendTypo}
\title{LegendTypo}
\usage{
LegendTypo(pos = "bottomleft", legTitle = "Title of the legend",
  legTitleCex = 0.8, legValuesCex = 0.6, categ, cols, nodata = TRUE,
  nodatalabel = "No data", frame = FALSE, symbol = "box")
}
\arguments{
\item{pos}{bottomleft, topleft, topright, bottomright, left, right, top, bottom, middle}

\item{legTitle}{Title of the legend}

\item{legTitleCex}{legTitleCex}

\item{legValuesCex}{legValuesCex}

\item{categ}{factor of categories}

\item{cols}{vector of colors}

\item{nodata}{if TRUE, a white box 'no data' is drawn.}

\item{nodatalabel}{label for no data value. Default = "No data"}

\item{frame}{if TRUE, a frame is drawn}

\item{symbol}{'line' or 'box'}
}
\value{
plot
}
\description{
legend for quali layers
}
\details{
plot a legend
}
\examples{
data("nuts2006")
plot(nuts0.spdf)
rect(par()$usr[1], par()$usr[3], par()$usr[2], par()$usr[4], border = "black")
LegendTypo(pos = "topright", legTitle = "Title of the legend, year",
           legTitleCex = 0.8, legValuesCex = 0.6,
           categ = c("John coltrane","Miles Davis","Maceo Parker",
           "Herbie\\nHancock","Julien Lourau"),
           cols = c("red","blue","green","yellow","purple"),
           nodatalabel = "Absence\\nd'information",nodata = TRUE,
           frame=TRUE,symbol="box")

LegendTypo(pos = "right", legTitle = "Title of the legend, year",
           legTitleCex = 0.8, legValuesCex = 0.6,
           categ = c("John coltrane","Miles Davis","Maceo Parker",
           "Herbie\\nHancock","Julien Lourau"),
           cols = c("red","blue","green","yellow","purple"),
           nodatalabel = "Absence\\nd'information",nodata = TRUE,
           frame=TRUE, symbol="line")
}

