% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{maphist}
\alias{maphist}
\title{Map Hist}
\usage{
maphist(v, p, cols, bks)
}
\arguments{
\item{v}{position}

\item{p}{c(px, py)}

\item{cols}{cols}

\item{bks}{bks}
}
\description{
Plot hist.
}
\examples{
library(cartography)
library(sf)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package="cartography"))
# Population density
mtq$POPDENS <- 1e6 * mtq$POP / st_area(x = mtq)
par(mar = c(0,0,0,0))
bks <- getBreaks(mtq$POPDENS, method = "quantile", nclass = 40)
cols <- carto.pal(pal1 = "wine.pal", 6)
choroLayer(x = mtq, var = "POPDENS", breaks = bks, col = cols,
           border = NA, legend.pos = "n")
pos <- "bottom"
fdim <- legendChoro(pos = pos, title.txt = "super long", breaks = bks,
                    col = cols, nodata = FALSE, frame = TRUE, hist = TRUE)
hist <- c(fdim[1],fdim[2],0,0)
if(substr(pos, 1,3)=="top"){
  hist[4] <- fdim[3]
  hist[3] <- fdim[3] - diff(fdim[1:2])
}else{
  hist[3] <- fdim[3]
  hist[4] <- fdim[3] + diff(fdim[1:2])
  fdim[3:4] <- fdim[3:4] + diff(fdim[1:2])
  pos <- fdim[c(1,3)]
}

maphist(v = mtq$POPDENS, p = hist, cols = cols, bks = bks)
legendChoro(pos = pos, title.txt = "super long", breaks = bks, col = cols,
           nodata = FALSE, frame = TRUE, hist = FALSE)
}
