% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts2.spdf}
\alias{nuts2.spdf}
\title{Nuts2 Regions}
\format{SpatialPolygonsDataFrame.}
\source{
UMS RIATE - \url{http://www.ums-riate.fr/Webriate/?page_id=153}
}
\description{
Delineations of EU administrative units (level 2, 2006 version).
}
\details{
This SpatialPolygonsDataFrame can be used with the nuts2.df data frame
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique nuts id (character)}
}}

