% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts3.df}
\alias{nuts3.df}
\title{Nuts3 Dataset}
\source{
UMS RIATE - \url{http://www.ums-riate.fr/Webriate/?page_id=151}. 
Data extraction: 2011; data validity: 2008.
}
\description{
This dataset contains some socio-economic data
}
\details{
This data frame can be used with the SpatialPolygonsDataFrame 
nuts3.spdf
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique nuts id (character)}

\item{\code{birth_2008}}{Number of birth in 2008 (live birth) (numeric)}

\item{\code{death_2008}}{Number of death in 2008 (death) (numeric)}

\item{\code{gdppps1999}}{Gross domestic product (Purchasing Power Standards) in 1999 
(million euros) (numeric)}

\item{\code{gdppps2008}}{Gross domestic product (Purchasing Power Standards) in 2008 
(million euros) (numeric)}

\item{\code{pop1999}}{Total population in 1999 (inhabitants) (numeric)}

\item{\code{pop2008}}{Total population in 2008 (inhabitants) (numeric)}
}}

