% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propLinkLayer.R
\name{propLinkLayer}
\alias{propLinkLayer}
\title{Proportional Links Layer}
\usage{
propLinkLayer(sldf, df, sldfid = NULL, sldfids, sldfide, dfid = NULL, dfids,
  dfide, var, maxlwd = 40, col, legend.pos = "bottomleft",
  legend.title.txt = var, legend.title.cex = 0.8, legend.values.cex = 0.6,
  legend.frame = FALSE, legend.values.rnd = 0, add = TRUE)
}
\arguments{
\item{sldf}{SpatialLinesDataFrame; a link layer.}

\item{df}{DataFrame with identifiers and a variable.}

\item{sldfid}{Unique identifier in sldf (sldfids, sldfide, dfids and dfide are not used).}

\item{sldfids}{Identifier of starting points in sldf (sldfid and dfid are not used).}

\item{sldfide}{Identifier of ending points in sldf (sldfid and dfid are not used).}

\item{dfid}{Unique identifier in df (sldfids, sldfide, dfids and dfide are not used).}

\item{dfids}{Identifier of starting points in df (sldfid and dfid are not used).}

\item{dfide}{Identifier of ending points in df (sldfid and dfid are not used).}

\item{var}{Name of the variable used to plot the links}

\item{maxlwd}{Maximum size of the links.}

\item{col}{Color of the links.}

\item{legend.pos}{character; position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright". If
legend.pos is "n" then the legend is not plotted.}

\item{legend.title.txt}{character; title of the legend.}

\item{legend.title.cex}{numeric; size of the legend title.}

\item{legend.values.cex}{numeric; size of the values in the legend.}

\item{legend.frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{legend.values.rnd}{numeric; number of decimal places of the values
displayed in the legend.}

\item{add}{boolean; whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot a layer of proportionnal links
}
\details{
Unlike most of cartography functions, identifiers fields are mandatory.
}
\examples{
data("nuts2006")
# Create a link layer
twincities.spdf <- getLinkLayer(spdf = nuts2.spdf, df = twincities[,1:2])
# plot the links
plot(nuts0.spdf, col = "grey60",border = "grey20")
propLinkLayer(sldf = twincities.spdf, df = twincities[twincities$fij>=5,],maxlwd = 10,
              sldfids = "i", sldfide = "j",
              dfids = "i", dfide = "j",legend.pos = "topright",
              var = "fij",
              col = "#92000090", add = TRUE)
}

