% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linkLayer.R
\name{propLinkLayer}
\alias{propLinkLayer}
\title{Layer of Proportional Links}
\usage{
propLinkLayer(spdf, df, spdfid = NULL, dfidi = NULL, dfidj = NULL,
  var1 = NULL, var2 = NULL, maxlwd = 40)
}
\arguments{
\item{spdf}{Spatial*DataFrame}

\item{df}{DataFrame with Ids and Labels}

\item{spdfid}{Ids of the obj Spatial*DataFrame}

\item{dfidi}{Ids of origins}

\item{dfidj}{Ids of destinations}

\item{var1}{name of the variable corresponding to lwd of the links}

\item{var2}{name of the variable corresponding to colors}

\item{maxlwd}{max lwd of the links}
}
\value{
a plot
}
\description{
Plot a layer of links
}

