% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/propSymbolsLayer.R
\name{propSymbolsLayer}
\alias{propSymbolsLayer}
\title{Proportional Symbols Layer}
\usage{
propSymbolsLayer(spdf, df, spdfid = NULL, dfid = NULL, var,
  symbols = "circles", col = "#E84923", col2 = "#7DC437",
  breakval = NULL, k = 0.02, fixmax = NULL, legend.pos = "bottomleft",
  legend.title.txt = var, legend.title.cex = 0.8, legend.values.cex = 0.6,
  legend.style = "a", legend.frame = FALSE, legend.values.rnd = 0,
  add = TRUE)
}
\arguments{
\item{spdf}{Spatial*DataFrame; if \code{spdf} is a SpatialPolygonsDataFrame
symbols are plotted on centroids.}

\item{df}{data.frame; \code{df} contains the values to plot.}

\item{spdfid}{character; id field in \code{spdf}, default to the first column
of the \code{spdf} data.frame. (optional)}

\item{dfid}{character; id field in \code{df}, default to the first column
of \code{df}. (optional)}

\item{var}{character; name of the numeric field in \code{df} to plot.}

\item{symbols}{character; type of symbols, one of "circles", "squares" or "height").}

\item{col}{character; color of symbols.}

\item{col2}{character; second color of symbols (see Details).}

\item{breakval}{numeric; breaking value (see Details).}

\item{k}{numeric; share of the map occupied by the biggest symbol.}

\item{fixmax}{numeric; value of the biggest symbol. (optional)}

\item{legend.pos}{character; position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{legend.title.txt}{character; title of the legend.}

\item{legend.title.cex}{numeric; size of the legend title.}

\item{legend.values.cex}{numeric; size of the values in the legend.}

\item{legend.style}{character; either "a" or "b". The legend has two display
styles.}

\item{legend.frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{legend.values.rnd}{numeric; number of decimal places of the values in
the legend.}

\item{add}{boolean; whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\value{
A plot is returned.
}
\description{
Plot a proportional symbols layer. Various symbols are availables.
}
\details{
The (\code{breakval}) parameter allows to plot symbols of two
colors: the first color (\code{col}) for values superior or equal to breakval,
second color (\code{col2}) for values inferior to breakval.
}
\examples{
data("nuts2006")

# Layout plot
layoutLayer(title = "Countries Population in Europe",
            sources = "UMS RIATE, 2015",
            author = "UMS RIATE",
            scale = NULL,
            frame = TRUE,
            col = "black",
            coltitle = "white",
            bg = "#D9F5FF",
            south = TRUE,
            extent = nuts0.spdf)

# Countries plot
plot(nuts0.spdf, col = "grey60",border = "grey20", add=TRUE)

# Population plot on proportional symbols
propSymbolsLayer(spdf = nuts0.spdf, df = nuts0.df,
                 var = "pop2008", k = 0.01,
                 symbols = "squares", col =  "#920000",
                 legend.pos = "right",
                 legend.title.txt = "Total\\npopulation (2008)",
                 legend.style = "a")
}

