% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propSymbolsTypoLayer.R
\name{propSymbolsTypoLayer}
\alias{propSymbolsTypoLayer}
\title{Proportional Symbols Typo Layer}
\usage{
propSymbolsTypoLayer(spdf, df, spdfid = NULL, dfid = NULL, var, var2,
  symbols = "circles", col = NULL, k = 0.02, fixmax = NULL,
  legend.var.pos = "bottomleft", legend.var2.pos = "bottomleft",
  legend.title.var.txt = var, legend.title.var2.txt = var2,
  legend.title.cex = 0.8, legend.values.cex = 0.6, legend.style = "a",
  legend.frame = FALSE, legend.values.rnd = 0, add = TRUE)
}
\arguments{
\item{spdf}{Spatial*DataFrame; if \code{spdf} is a SpatialPolygonsDataFrame
symbols are plotted on centroids.}

\item{df}{data.frame; \code{df} contains the values to plot.}

\item{spdfid}{character; id field in \code{spdf}, default to the first column
of the \code{spdf} data.frame. (optional)}

\item{dfid}{character; id field in \code{df}, default to the first column
of \code{df}. (optional)}

\item{var}{character; name of the numeric field in \code{df} to plot.}

\item{var2}{character; name of the factor field in \code{df} to plot.}

\item{symbols}{character; type of symbols, one of "circles", "squares" or "height").}

\item{col}{character; a vector of colors (hues).}

\item{k}{numeric; share of the map occupied by the biggest symbol.}

\item{fixmax}{numeric; value of the biggest symbol. (optional)}

\item{legend.var.pos}{character; position of the legend var, one of "topleft", "top",}

\item{legend.var2.pos}{character; position of the legend var2, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{legend.title.var.txt}{character; title of the legend (numeric data).}

\item{legend.title.var2.txt}{character; title of the legend (factor data).}

\item{legend.title.cex}{numeric; size of the legend title.}

\item{legend.values.cex}{numeric; size of the values in the legend.}

\item{legend.style}{character; either "a" or "b". The legend has two display
styles.}

\item{legend.frame}{boolean; whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{legend.values.rnd}{numeric; number of decimal places of the values in
the legend.}

\item{add}{boolean; whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\value{
A plot is returned.
}
\description{
Plot a proportional symbols layer with a color based on a qualitative data. Various symbols are availables.
}
\examples{
data("nuts2006")
# Layout plot
layoutLayer(title = "Countries Population in Europe",
           sources = "UMS RIATE, 2015",
           author = "UMS RIATE",
           scale = NULL,
           frame = TRUE,
           col = "black",
          coltitle = "white",
           bg = "#D9F5FF",
           south = TRUE,
           extent = nuts0.spdf)
#Countries plot
plot(nuts0.spdf, col = "grey60",border = "grey20", add=TRUE)
#Population plot on proportional symbols
#colours <- palette(rainbow(length(nuts0.spdf)))
#propSymbolsTypoLayer(spdf = nuts0.spdf, df = nuts0.df,
#                     var = "pop2008", var2="id", k = 0.01,
#                     symbols = "circles", col =  colours,
#                     legend.var.pos = "topleft", legend.var2.pos = "topright",
#                     legend.title.var.txt = "Total\\npopulation (2008)",
#                     legend.title.var2.txt = "Types de trucs",
#                     legend.style = "b")
}

