% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesLayer.R
\name{tilesLayer}
\alias{tilesLayer}
\title{Plot Tiles from Open Map Servers}
\usage{
tilesLayer(x, add = FALSE)
}
\arguments{
\item{x}{a RasterBrick object; the \link{getTiles} function outputs these 
objects.}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}
}
\description{
Plot tiles from open map servers.
}
\note{
This function is a wrapper for plotRGB from the 
raster package.
}
\seealso{
\link{getTiles}
}

