% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesLayer.R
\name{tilesLayer}
\alias{tilesLayer}
\title{Plot Tiles}
\usage{
tilesLayer(x, add = FALSE, ...)
}
\arguments{
\item{x}{a RasterBrick object; \link{getTiles} 
function outputs these objects.}

\item{add}{whether to add the layer to an existing plot (TRUE) or 
not (FALSE).}

\item{...}{bgalpha, interpolate, or other arguments passed to be passed to
\code{\link[raster:plotRGB]{plotRGB}}}
}
\description{
Plot a raster object over a map. It can be used to plot tiles 
from getTiles.
}
\examples{
library(sf)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package = "cartography"))

}
\seealso{
\link[maptiles]{plot_tiles}
}
\keyword{internal}
