% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmLayer.R
\name{tilesLayer}
\alias{tilesLayer}
\title{Plot Tiles From Open Map Servers}
\usage{
tilesLayer(x, add = FALSE)
}
\arguments{
\item{x}{an OpenStreetMap object; \code{\link{getTiles}} outputs these objects.}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot Tiles From Open Map Servers
}
\details{
This function is a wrapper for \code{plot.OpenStreetMap} from
OpenStreetMap package.
}
\examples{
data("nuts2006")
# Do not run:
# Download the tiles
#StamWatCol <- getTiles(spdf = nuts0.spdf, type = "stamen-watercolor")
#Display the tiles Layer
#tilesLayer(x = StamWatCol)
#Display countries boundaries
#plot(nuts0.spdf, add = TRUE)
}

