% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/typoLayer.R
\name{typoLayer}
\alias{typoLayer}
\title{Topology Layer}
\usage{
typoLayer(spdf, df, spdfid = NULL, dfid = NULL, var, col = NULL,
  border = NA, lwd = 1, legend.pos = "bottomleft",
  legend.title.txt = var, legend.title.cex = 0.8, legend.values.cex = 0.6,
  legend.frame = FALSE, legend.nodata = "no data", add = TRUE)
}
\arguments{
\item{spdf}{a SpatialPolygonsDataFrame.}

\item{df}{a data.frame that contains the values to plot.}

\item{spdfid}{identifier field in spdf, default to the first column
of the spdf data.frame. (optional)}

\item{dfid}{identifier field in df, default to the first column
of df. (optional)}

\item{var}{name of the factor field in df to plot.}

\item{col}{a vector of colors. Note that if breaks is specified there must be one less
colors specified than the number of break.}

\item{border}{color of polygon borders.}

\item{lwd}{borders width.}

\item{legend.pos}{position of the legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright". If
legend.pos is "n" then the legend is not plotted.}

\item{legend.title.txt}{title of the legend.}

\item{legend.title.cex}{size of the legend title.}

\item{legend.values.cex}{size of the values in the legend.}

\item{legend.frame}{whether to add a frame to the legend (TRUE) or
not (FALSE).}

\item{legend.nodata}{no data label.}

\item{add}{whether to add the layer to an existing plot (TRUE) or
not (FALSE).}
}
\description{
Plot a typology layer.
}
\examples{
data(nuts2006)
nuts0.df$typo <- substr(nuts0.df$id,2,2)
nuts0.df$typo <- c(rep("A",10),rep("B",10),rep("C",10),rep("D",4))
colours <- c("red","green","blue","yellow")

typoLayer(spdf = nuts0.spdf, df = nuts0.df, var = "typo", col=colours,
          legend.pos = "topright", border="white",
          legend.title.txt = "Typologie",add=FALSE
)
}
\references{
Herbert A. Sturges, «
\emph{The Choice of a Class Interval }», Journal of the American Statistical Association, vol. 21, n° 153, mars 1926, p. 65-66.
}

